Module icu::calendar

Contains the core types used by ICU4X for dealing with dates, times, and custom calendars.

The types module has a lot of common types for dealing with dates and times.

Calendar is a trait that allows one to define custom calendars, and Date can represent dates for arbitrary calendars.

The iso and gregorian modules contain implementations for the ISO and Gregorian calendars respectively.

Modules

any_calendar

Module for working with multiple calendars at once

arithmetic

Miscellaneous helper functions.

buddhist

This module contains types and implementations for the Buddhist calendar.

coptic

This module contains types and implementations for the Coptic calendar.

ethiopic

This module contains types and implementations for the Ethiopic calendar.

gregorian

This module contains types and implementations for the Gregorian calendar.

indian

This module contains types and implementations for the Indian national calendar.

iso

This module contains types and implementations for the ISO calendar.

japanese

This module contains types and implementations for the Japanese calendar

julian

This module contains types and implementations for the Julian calendar.

provider

Data provider struct definitions for this ICU4X component.

types

This module contains various types used by icu_calendar and icu_datetime

Structs

ArithmeticDate

Date

A date for a given calendar

DateDuration

A duration between two dates

DateTime

A date+time for a given calendar

Gregorian

The Gregorian Calendar

Iso

The ISO Calendar

Enums

DateDurationUnit

A “duration unit” used to specify the minimum or maximum duration of time to care about

DateTimeError

A list of possible error outcomes for working with various inputs to DateTime inputs and operations.

Traits

AsCalendar

Types that contain a calendar

Calendar

A calendar implementation

CalendarArithmetic