icu::calendar::cal

Enum AnyCalendar

#[non_exhaustive]
pub enum AnyCalendar {

    Buddhist(Buddhist),
    Chinese(EastAsianTraditional<China>),
    Coptic(Coptic),
    Dangi(EastAsianTraditional<Korea>),
    Ethiopian(Ethiopian),
    Gregorian(Gregorian),
    Hebrew(Hebrew),
    Indian(Indian),
    HijriTabular(Hijri<TabularAlgorithm>),
    HijriSimulated(Hijri<AstronomicalSimulation>),
    HijriUmmAlQura(Hijri<UmmAlQura>),
    Iso(Iso),
    Japanese(Japanese),
    JapaneseExtended(JapaneseExtended),
    Persian(Persian),
    Roc(Roc),

}

This is a calendar that encompasses all formattable calendars supported by this crate

This allows for the construction of Date objects that have their calendar known at runtime.

This can be constructed by calling .into() on a concrete calendar type if the calendar type is known at compile time. When the type is known at runtime, the AnyCalendar::new() and sibling methods may be used.

Date can also be converted to AnyCalendar-compatible ones via Date::to_any().

There are many ways of constructing an AnyCalendarโd date:

use icu::calendar::{AnyCalendar, AnyCalendarKind, Date, cal::{Japanese, Gregorian}, types::MonthCode};
use icu::locale::locale;
use tinystr::tinystr;

let locale = locale!("en-u-ca-japanese"); // English with the Japanese calendar

let calendar = AnyCalendar::new(AnyCalendarKind::new(locale.into()));

// This is a Date<AnyCalendar>
let any_japanese_date = Date::try_new_gregorian(2020, 9, 1)
    .expect("Failed to construct Gregorian Date.")
    .to_calendar(calendar)
    .to_any();

// Construct a date in the appropriate typed calendar and convert
let japanese_calendar = Japanese::new();
let japanese_date = Date::try_new_japanese_with_calendar("reiwa", 2, 9, 1,
                                                        japanese_calendar).unwrap();
assert_eq!(japanese_date.to_any(), any_japanese_date);

// this is also Date<AnyCalendar>, but it uses a different calendar
let any_gregorian_date = any_japanese_date.to_calendar(Gregorian).to_any();

// Date<AnyCalendar> does not have a total order
assert!(any_gregorian_date <= any_gregorian_date);
assert!(any_japanese_date <= any_japanese_date);
assert!(!(any_gregorian_date <= any_japanese_date) && !(any_japanese_date <= any_gregorian_date));

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Buddhist(Buddhist)

A Buddhist calendar

Chinese(EastAsianTraditional<China>)

A Chinese calendar

Coptic(Coptic)

A Coptic calendar

Dangi(EastAsianTraditional<Korea>)

A Dangi calendar

Ethiopian(Ethiopian)

An Ethiopian calendar

Gregorian(Gregorian)

A Gregorian calendar

Hebrew(Hebrew)

A Hebrew calendar

Indian(Indian)

An Indian calendar

HijriTabular(Hijri<TabularAlgorithm>)

A HijriTabular calendar

HijriSimulated(Hijri<AstronomicalSimulation>)

A HijriSimulated calendar

HijriUmmAlQura(Hijri<UmmAlQura>)

A HijriUmmAlQura calendar

Iso(Iso)

An Iso calendar

Japanese(Japanese)

A Japanese calendar

JapaneseExtended(JapaneseExtended)

A JapaneseExtended calendar

Persian(Persian)

A Persian calendar

Roc(Roc)

A Roc calendar

Implementations

impl AnyCalendar

pub const fn new(kind: AnyCalendarKind) -> AnyCalendar

Constructs an AnyCalendar for a given calendar kind from compiled data.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

pub fn try_new_with_buffer_provider<P>( provider: &P, kind: AnyCalendarKind, ) -> Result<AnyCalendar, DataError>
where P: BufferProvider + ?Sized,

A version of Self::new that uses custom data provided by a BufferProvider.

โจ Enabled with the serde feature.

๐ Help choosing a constructor

pub fn try_new_unstable<P>( provider: &P, kind: AnyCalendarKind, ) -> Result<AnyCalendar, DataError>
where P: DataProvider<CalendarJapaneseModernV1> + DataProvider<CalendarJapaneseExtendedV1> + ?Sized,

A version of Self::new that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

โ๏ธ The bounds on provider may change over time, including in SemVer minor releases.

pub fn kind(&self) -> AnyCalendarKind

The AnyCalendarKind corresponding to the calendar this contains