icu::calendar::cal::scaffold

Trait UnstableSealed

pub trait UnstableSealed { }

Trait marking other traits that are considered unstable and should not generally be implemented outside of the calendar crate.

๐ง This trait is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. Do not implement this trait in userland unless you are prepared for things to occasionally break.