icu::calendar::cal

Struct HijriUmmAlQura

#[non_exhaustive]
pub struct HijriUmmAlQura;

The Umm al-Qura Hijri Calendar

This calendar is the official calendar in Saudi Arabia.

Era codes

This calendar uses two era codes: ah, and bh, corresponding to the Anno Hegirae and Before Hijrah eras

Month codes

This calendar is a pure lunar calendar with no leap months. It uses month codes "M01" - "M12".

Implementations

impl HijriUmmAlQura

pub const fn new() -> HijriUmmAlQura

Creates a new HijriUmmAlQura.