icu::calendar::cal

Struct Iso

pub struct Iso;

The ISO-8601 Calendar

This calendar is identical to the Gregorian calendar, except that it uses a single default era instead of bce and ce.

This corresponds to the "iso8601" CLDR calendar.

Era codes

This calendar uses a single era: default

Implementations

impl Iso

pub fn new() -> Iso

Construct a new ISO Calendar