Struct icu::calendar::islamic::IslamicTabular

pub struct IslamicTabular;

A Tabular version of the Arithmetical Islamic Calendar

Era codes

This calendar supports a single era code, Anno Mundi, with code "ah"

Month codes

This calendar is a pure lunar calendar with no leap months. It uses month codes "M01" - "M12".

Implementations

impl IslamicTabular

pub fn new() -> IslamicTabular

Construct a new IslamicTabular

pub fn new_always_calculating() -> IslamicTabular

๐Deprecated: Precomputation not needed for this calendar

Construct a new IslamicTabular (deprecated: we will not add precomputation to this calendar)