Struct icu::calendar::iso::Iso

pub struct Iso;

The ISO Calendar

The ISO Calendar is a standardized solar calendar with twelve months. It is identical to the Gregorian calendar, except it uses negative years for years before 1 CE, and may have differing formatting data for a given locale.

This type can be used with Date or DateTime to represent dates in this calendar.

Era codes

This calendar supports one era, "default"

Implementations

impl Iso

pub fn new() -> Iso

Construct a new ISO Calendar