Struct icu::calendar::japanese::Japanese

pub struct Japanese { /* private fields */ }

The Japanese Calendar (with modern eras only)

The Japanese calendar is a solar calendar used in Japan, with twelve months. The months and days are identical to that of the Gregorian calendar, however the years are counted differently using the Japanese era system.

This calendar only contains eras after Meiji, for all historical eras, check out JapaneseExtended.

This type can be used with Date or DateTime to represent dates in this calendar.

Era codes

This calendar currently supports seven era codes. It supports the five post-Meiji eras ("meiji", "taisho", "showa", "heisei", "reiwa"), as well as using the Gregorian "bce" and "ce" for dates before the Meiji era.

Future eras will also be added to this type when they are decided.

These eras are loaded from data, requiring a data provider capable of providing JapaneseErasV1Marker data (calendar/japanese@1).

Implementations

impl Japanese

pub fn try_new_unstable<D>(data_provider: &D) -> Result<Japanese, CalendarError>where D: DataProvider<JapaneseErasV1Marker> + ?Sized,

Creates a new Japanese from locale data using only modern eras (post-meiji).

๐ Help choosing a constructor

โ๏ธ The bounds on this function may change over time, including in SemVer minor releases.

pub fn try_new_with_any_provider( provider: &impl AnyProvider ) -> Result<Japanese, CalendarError>

Creates a new instance using an AnyProvider.

For details on the behavior of this function, see: Self::try_new_unstable

๐ Help choosing a constructor

pub fn try_new_with_buffer_provider( provider: &impl BufferProvider ) -> Result<Japanese, CalendarError>

โจ Enabled with the "serde" feature.

Creates a new instance using a BufferProvider.

For details on the behavior of this function, see: Self::try_new_unstable

๐ Help choosing a constructor