Struct icu::calendar::japanese::JapaneseExtended

pub struct JapaneseExtended(_);

The Japanese Calendar (with historical eras)

The Japanese calendar is a solar calendar used in Japan, with twelve months. The months and days are identical to that of the Gregorian calendar, however the years are counted differently using the Japanese era system.

This type can be used with Date or DateTime to represent dates in this calendar.

Era codes

This calendar supports a large number of era codes. It supports the five post-Meiji eras ("meiji", "taisho", "showa", "heisei", "reiwa"). Pre-Meiji eras are represented with their names converted to lowercase ascii and followed by their start year. E.g. the “Ten’ō” era (781 - 782 CE) has the code "teno-781". The Gregorian "bce" and "ce" eras are used for dates before the first known era era.

These eras are loaded from data, requiring a data provider capable of providing JapaneseExtendedErasV1Marker data (calendar/japanext@1).

Implementations

impl JapaneseExtended

pub fn try_new_unstable<D>(
data_provider: &D
) -> Result<JapaneseExtended, DataError> where
D: DataProvider<JapaneseExtendedErasV1Marker> + ?Sized,

Creates a new Japanese from locale data using all eras (including pre-meiji).

📚 Help choosing a constructor

⚠️ The bounds on this function may change over time, including in SemVer minor releases.

pub fn try_new_with_any_provider(
provider: &impl AnyProvider
) -> Result<JapaneseExtended, DataError>

Creates a new instance using an AnyProvider.

For details on the behavior of this function, see: Self::try_new_unstable

📚 Help choosing a constructor

pub fn try_new_with_buffer_provider(
provider: &impl BufferProvider
) -> Result<JapaneseExtended, DataError>

✨ Enabled with the "serde" feature.

Creates a new instance using a BufferProvider.

For details on the behavior of this function, see: Self::try_new_unstable

📚 Help choosing a constructor