icu::calendar

Module provider

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Modules

chinese_based

🚧 [Unstable] Data provider struct definitions for chinese-based calendars.

hijri

🚧 [Unstable] Data provider struct definitions for chinese-based calendars.

Structs

Baked

Baked data

CalendarChineseV1

Precomputed data for the Chinese calendar

CalendarDangiV1

Precomputed data for the Dangi calendar

CalendarHijriSimulatedMeccaV1

Precomputed data for the Hijri obsevational calendar

CalendarJapaneseExtendedV1

Full Japanese era names

CalendarJapaneseModernV1

Modern Japanese era names

CalendarWeekV1

Week information

EraStartDate

The date at which an era started

EraStartDateULE

ULE type for EraStartDate

JapaneseEras

A data structure containing the necessary era data for constructing a Japanese calendar object

WeekData

An ICU4X mapping to a subset of CLDR weekData. See CLDR-JSON’s weekData.json for more context.

WeekdaySet

Bitset representing weekdays.