Struct icu::calendar::DateTime

pub struct DateTime<A> where
    A: AsCalendar,  {
    pub date: Date<A>,
    pub time: Time,
}

A date+time for a given calendar.

This can work with wrappers around Calendar types, e.g. Rc<C>, via the AsCalendar trait, much like Date.

use icu::calendar::DateTime;

// Example: Construction of ISO datetime from integers.
let datetime_iso = DateTime::new_iso_datetime(1970, 1, 2, 13, 1, 0)
    .expect("Failed to initialize ISO DateTime instance.");

assert_eq!(datetime_iso.date.year().number, 1970);
assert_eq!(datetime_iso.date.month().ordinal, 1);
assert_eq!(datetime_iso.date.day_of_month().0, 2);
assert_eq!(datetime_iso.time.hour.number(), 13);
assert_eq!(datetime_iso.time.minute.number(), 1);
assert_eq!(datetime_iso.time.second.number(), 0);

Fields
date: Date<A>

The date

time: Time

The time

Implementations

impl DateTime<Buddhist>

pub fn new_buddhist_datetime(
year: i32,
month: u8,
day: u8,
hour: u8,
minute: u8,
second: u8
) -> Result<DateTime<Buddhist>, DateTimeError>

Construct a new Buddhist datetime from integers.

Years are specified as BE years.

use icu::calendar::DateTime;

let datetime_buddhist = DateTime::new_buddhist_datetime(1970, 1, 2, 13, 1, 0)
    .expect("Failed to initialize Buddhist DateTime instance.");

assert_eq!(datetime_buddhist.date.year().number, 1970);
assert_eq!(datetime_buddhist.date.month().ordinal, 1);
assert_eq!(datetime_buddhist.date.day_of_month().0, 2);
assert_eq!(datetime_buddhist.time.hour.number(), 13);
assert_eq!(datetime_buddhist.time.minute.number(), 1);
assert_eq!(datetime_buddhist.time.second.number(), 0);

impl DateTime<Coptic>

pub fn new_coptic_datetime(
year: i32,
month: u8,
day: u8,
hour: u8,
minute: u8,
second: u8
) -> Result<DateTime<Coptic>, DateTimeError>

Construct a new Coptic datetime from integers.

Negative years are in the B.D. era, starting with 0 = 1 B.D.

use icu::calendar::DateTime;

let datetime_coptic = DateTime::new_coptic_datetime(1686, 5, 6, 13, 1, 0)
    .expect("Failed to initialize Coptic DateTime instance.");

assert_eq!(datetime_coptic.date.year().number, 1686);
assert_eq!(datetime_coptic.date.month().ordinal, 5);
assert_eq!(datetime_coptic.date.day_of_month().0, 6);
assert_eq!(datetime_coptic.time.hour.number(), 13);
assert_eq!(datetime_coptic.time.minute.number(), 1);
assert_eq!(datetime_coptic.time.second.number(), 0);

impl<A> DateTime<A> where
A: AsCalendar,

pub fn new(date: Date<A>, time: Time) -> DateTime<A>

Construct a DateTime for a given Date and Time

pub fn new_from_codes(
era: Era,
year: i32,
month_code: MonthCode,
day: u8,
time: Time,
calendar: A
) -> Result<DateTime<A>, DateTimeError>

Construct a datetime from from era/month codes and fields, and some calendar representation

pub fn new_from_iso(iso: DateTime<Iso>, calendar: A) -> DateTime<A>

Construct a DateTime from an ISO datetime and some calendar representation

pub fn to_iso(&self) -> DateTime<Iso>

Convert the DateTime to an ISO DateTime

pub fn to_calendar<A2>(&self, calendar: A2) -> DateTime<A2> where
A2: AsCalendar,

Convert the DateTime to a DateTime in a different calendar

impl<C, A> DateTime<A> where
C: IntoAnyCalendar,
A: AsCalendar<Calendar = C>,

pub fn to_any(&self) -> DateTime<AnyCalendar>

Type-erase the date, converting it to a date for AnyCalendar

impl DateTime<Ethiopian>

pub fn new_ethiopian_datetime(
era_style: EthiopianEraStyle,
year: i32,
month: u8,
day: u8,
hour: u8,
minute: u8,
second: u8
) -> Result<DateTime<Ethiopian>, DateTimeError>

Construct a new Ethiopian datetime from integers.

For the Amete Mihret era style, negative years work with year 0 as 1 pre-Incarnation, year -1 as 2 pre-Incarnation, and so on.

use icu::calendar::DateTime;
use icu::calendar::ethiopian::EthiopianEraStyle;

let datetime_ethiopian = DateTime::new_ethiopian_datetime(EthiopianEraStyle::AmeteMihret, 2014, 8, 25, 13, 1, 0)
    .expect("Failed to initialize Ethiopian DateTime instance.");

assert_eq!(datetime_ethiopian.date.year().number, 2014);
assert_eq!(datetime_ethiopian.date.month().ordinal, 8);
assert_eq!(datetime_ethiopian.date.day_of_month().0, 25);
assert_eq!(datetime_ethiopian.time.hour.number(), 13);
assert_eq!(datetime_ethiopian.time.minute.number(), 1);
assert_eq!(datetime_ethiopian.time.second.number(), 0);

impl DateTime<Gregorian>

pub fn new_gregorian_datetime(
year: i32,
month: u8,
day: u8,
hour: u8,
minute: u8,
second: u8
) -> Result<DateTime<Gregorian>, DateTimeError>

Construct a new Gregorian datetime from integers.

Years are specified as ISO years.

use icu::calendar::DateTime;

let datetime_gregorian = DateTime::new_gregorian_datetime(1970, 1, 2, 13, 1, 0)
    .expect("Failed to initialize Gregorian DateTime instance.");

assert_eq!(datetime_gregorian.date.year().number, 1970);
assert_eq!(datetime_gregorian.date.month().ordinal, 1);
assert_eq!(datetime_gregorian.date.day_of_month().0, 2);
assert_eq!(datetime_gregorian.time.hour.number(), 13);
assert_eq!(datetime_gregorian.time.minute.number(), 1);
assert_eq!(datetime_gregorian.time.second.number(), 0);

impl DateTime<Indian>

pub fn new_indian_datetime(
year: i32,
month: u8,
day: u8,
hour: u8,
minute: u8,
second: u8
) -> Result<DateTime<Indian>, DateTimeError>

Construct a new Indian datetime from integers, with year provided in the Śaka era.

use icu::calendar::DateTime;

let datetime_indian = DateTime::new_indian_datetime(1891, 10, 12, 13, 1, 0)
    .expect("Failed to initialize Indian DateTime instance.");

assert_eq!(datetime_indian.date.year().number, 1891);
assert_eq!(datetime_indian.date.month().ordinal, 10);
assert_eq!(datetime_indian.date.day_of_month().0, 12);
assert_eq!(datetime_indian.time.hour.number(), 13);
assert_eq!(datetime_indian.time.minute.number(), 1);
assert_eq!(datetime_indian.time.second.number(), 0);

impl DateTime<Iso>

pub fn new_iso_datetime(
year: i32,
month: u8,
day: u8,
hour: u8,
minute: u8,
second: u8
) -> Result<DateTime<Iso>, DateTimeError>

Construct a new ISO datetime from integers.

use icu::calendar::DateTime;

let datetime_iso = DateTime::new_iso_datetime(1970, 1, 2, 13, 1, 0)
    .expect("Failed to initialize ISO DateTime instance.");

assert_eq!(datetime_iso.date.year().number, 1970);
assert_eq!(datetime_iso.date.month().ordinal, 1);
assert_eq!(datetime_iso.date.day_of_month().0, 2);
assert_eq!(datetime_iso.time.hour.number(), 13);
assert_eq!(datetime_iso.time.minute.number(), 1);
assert_eq!(datetime_iso.time.second.number(), 0);

pub fn minutes_since_local_unix_epoch(&self) -> i32

Minute count representation of calendars starting from 00:00:00 on Jan 1st, 1970.

use icu::calendar::DateTime;

let today = DateTime::new_iso_datetime(2020, 2, 29, 0, 0, 0).unwrap();

assert_eq!(today.minutes_since_local_unix_epoch(), 26382240);
assert_eq!(
            DateTime::from_minutes_since_local_unix_epoch(26382240),
            Ok(today)
        );

let today = DateTime::new_iso_datetime(1970, 1, 1, 0, 0, 0).unwrap();

assert_eq!(today.minutes_since_local_unix_epoch(), 0);
assert_eq!(DateTime::from_minutes_since_local_unix_epoch(0), Ok(today));

pub fn from_minutes_since_local_unix_epoch(
minute: i32
) -> Result<DateTime<Iso>, DateTimeError>

Convert minute count since 00:00:00 on Jan 1st, 1970 to ISO Date.

use icu::calendar::DateTime;

let today = DateTime::new_iso_datetime(2020, 2, 29, 0, 0, 0).unwrap();

assert_eq!(today.minutes_since_local_unix_epoch(), 26382240);
assert_eq!(
            DateTime::from_minutes_since_local_unix_epoch(26382240),
            Ok(today)
        );

let today = DateTime::new_iso_datetime(1970, 1, 1, 0, 0, 0).unwrap();

assert_eq!(today.minutes_since_local_unix_epoch(), 0);
assert_eq!(DateTime::from_minutes_since_local_unix_epoch(0), Ok(today));

impl DateTime<Japanese>

pub fn new_japanese_datetime<A>(
era: Era,
year: i32,
month: u8,
day: u8,
hour: u8,
minute: u8,
second: u8,
japanese_calendar: A
) -> Result<DateTime<A>, DateTimeError> where
A: AsCalendar<Calendar = Japanese>,

Construct a new Japanese datetime from integers.

Years are specified in the era provided.

use icu::calendar::{types, DateTime};
use icu::calendar::japanese::Japanese;
use std::convert::TryFrom;
use tinystr::tinystr;

let provider = icu_testdata::get_provider();
let japanese_calendar = Japanese::try_new_with_buffer_provider(&provider).expect("Cannot load japanese data");

let era = types::Era(tinystr!(16, "heisei"));

let datetime = DateTime::new_japanese_datetime(era, 14, 1, 2, 13, 1, 0, japanese_calendar)
    .expect("Constructing a date should succeed");

assert_eq!(datetime.date.year().era, era);
assert_eq!(datetime.date.year().number, 14);
assert_eq!(datetime.date.month().ordinal, 1);
assert_eq!(datetime.date.day_of_month().0, 2);
assert_eq!(datetime.time.hour.number(), 13);
assert_eq!(datetime.time.minute.number(), 1);
assert_eq!(datetime.time.second.number(), 0);

impl DateTime<JapaneseExtended>

pub fn new_japanese_extended_datetime<A>(
era: Era,
year: i32,
month: u8,
day: u8,
hour: u8,
minute: u8,
second: u8,
japanext_calendar: A
) -> Result<DateTime<A>, DateTimeError> where
A: AsCalendar<Calendar = JapaneseExtended>,

Construct a new Japanese datetime from integers with all eras.

Years are specified in the era provided.

use icu::calendar::{types, DateTime};
use icu::calendar::japanese::JapaneseExtended;
use std::convert::TryFrom;
use tinystr::tinystr;

let provider = icu_testdata::get_provider();
let japanext_calendar = JapaneseExtended::try_new_with_buffer_provider(&provider).expect("Cannot load japanese data");

let era = types::Era(tinystr!(16, "kansei-1789"));

let datetime = DateTime::new_japanese_extended_datetime(era, 7, 1, 2, 13, 1, 0, japanext_calendar)
    .expect("Constructing a date should succeed");

assert_eq!(datetime.date.year().era, era);
assert_eq!(datetime.date.year().number, 7);
assert_eq!(datetime.date.month().ordinal, 1);
assert_eq!(datetime.date.day_of_month().0, 2);
assert_eq!(datetime.time.hour.number(), 13);
assert_eq!(datetime.time.minute.number(), 1);
assert_eq!(datetime.time.second.number(), 0);

impl DateTime<Julian>

pub fn new_julian_datetime(
year: i32,
month: u8,
day: u8,
hour: u8,
minute: u8,
second: u8
) -> Result<DateTime<Julian>, DateTimeError>

Construct a new Julian datetime from integers.

Zero and negative years are in BC, with year 0 = 1 BC

use icu::calendar::DateTime;

let datetime_julian = DateTime::new_julian_datetime(1969, 12, 20, 13, 1, 0)
    .expect("Failed to initialize Julian DateTime instance.");

assert_eq!(datetime_julian.date.year().number, 1969);
assert_eq!(datetime_julian.date.month().ordinal, 12);
assert_eq!(datetime_julian.date.day_of_month().0, 20);
assert_eq!(datetime_julian.time.hour.number(), 13);
assert_eq!(datetime_julian.time.minute.number(), 1);
assert_eq!(datetime_julian.time.second.number(), 0);