icu::calendar
Struct Gregorian

pub struct Gregorian;

The Gregorian Calendar

The Gregorian calendar is a solar calendar used by most of the world, with twelve months.

This type can be used with Date or DateTime to represent dates in this calendar.

Era codes

This calendar supports two era codes: "bce", and "ce", corresponding to the BCE and CE eras