Struct icu::calendar::types::FormattableYear

#[non_exhaustive]
pub struct FormattableYear {
    pub era: Era,
    pub number: i32,
    pub cyclic: Option<NonZero<u8>>,
    pub related_iso: Option<i32>,
}

Representation of a formattable year.

More fields may be added in the future for things like extended year

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

era: Era

The era containing the year.

This may not always be the canonical era for the calendar and could be an alias, for example all islamic calendars return islamic as the formattable era code which allows them to share data.

number: i32

The year number in the current era (usually 1-based).

cyclic: Option<NonZero<u8>>

The year in the current cycle for cyclic calendars (1-indexed) can be set to None for non-cyclic calendars

For chinese and dangi it will be a number between 1 and 60, for hypothetical other calendars it may be something else.

related_iso: Option<i32>

The related ISO year. This is normally the ISO (proleptic Gregorian) year having the greatest overlap with the calendar year. It is used in certain date formatting patterns.

Can be None if the calendar does not typically use related_iso (and CLDR does not contain patterns using it)

Implementations

impl FormattableYear

pub fn new( era: Era, number: i32, cyclic: Option<NonZero<u8>>, ) -> FormattableYear

Construct a new Year given an era and number

Other fields can be set mutably after construction as needed