icu::casemap::provider
Module exception_helpers

This module contains various types for the header part of casemapping exception data

This is both used in datagen to decode ICU4C’s data, and natively in ICU4X’s own data model.

ExceptionBits is the bag of bits associated with exceptions, and SlotPresence marks the presence or absence of various “slots” in a given exception.

The exceptions_builder module of this crate handles decoding ICU4C data using the exception header, and crate::provider::exceptions handles.

Structs

ExceptionBits

A bunch of bits associated with each exception.

ExceptionBitsULE

The bitflags on an exception header.

SlotPresence

Packed slot presence marker