Struct icu::casemap::provider::exception_helpers::ExceptionBitsULE

#[repr(transparent)]
pub struct ExceptionBitsULE(pub u8);

The bitflags on an exception header.

Format from icu4c, documented in casepropsbuilder.cpp, shifted 8 bits since ICU4C has this packed alongside a SlotPresence

           0  Double-width slots. If set, then each optional slot is stored as two
              elements of the array (high and low halves of 32-bit values) instead of
              a single element.
           1  Has no simple case folding, even if there is a simple lowercase mapping
          2  The value in the delta slot is negative
          3  Is case-sensitive (not exposed)
      4..5  Dot type
          6  Has conditional special casing
          7  Has conditional case folding

All bits are valid, though in ICU4X data bits 0 and 2 are not used

๐ง This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Tuple Fields
0: u8
Implementations

impl ExceptionBitsULE

pub fn double_width_slots(self) -> bool

Whether or not the slots are double-width.

Unused in ICU4X

pub fn no_simple_case_folding(self) -> bool

There is no simple casefolding, even if there is a simple lowercase mapping

pub fn negative_delta(self) -> bool

The delta stored in the Delta slot is negative

pub fn is_sensitive(self) -> bool

If the character is case sensitive

pub fn has_conditional_special(self) -> bool

If the character has conditional special casing

pub fn has_conditional_fold(self) -> bool

If the character has conditional case folding

pub fn dot_type(self) -> DotType

The dot type of the character