icu::casemap
Trait ClosureSink

pub trait ClosureSink {
    // Required methods
    fn add_char(&mut self, c: char);
    fn add_string(&mut self, string: &str);
}

An object that accepts characters and/or strings to be used with CaseMapCloser::add_string_case_closure_to() and CaseMapCloser::add_case_closure_to(). Usually this object will be some kind of set over codepoints and strings, or something that can be built into one.

Required Methods

fn add_char(&mut self, c: char)

Add a character to the set

fn add_string(&mut self, string: &str)

Add a string to the set