icu::collator

Module options

This module contains the types that are part of the API for setting the options for the collator.

Structs

CollatorOptions

Options settable by the user of the API.

ResolvedCollatorOptions

The resolved (actually used) options used by the collator.

Enums

AlternateHandling

What to do about characters whose comparison level can be varied dynamically.

CaseLevel

Whether to distinguish case in sorting, even for sorting levels higher than tertiary, without having to use tertiary level just to enable case level differences.

MaxVariable

What characters get shifted to the quaternary level with AlternateHandling::Shifted.

Strength

The collation strength that indicates how many levels to compare.