icu::collator

Module provider

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Structs

Baked

Baked data

CollationData

The main collation data either for the root or for a tailoring

CollationDiacritics

Secondary weights for the start of the Combining Diacritics block.

CollationDiacriticsV1

Data marker for collation diacritics data.

CollationJamo

CollationElement32s for the Hangul Jamo Unicode Block

CollationJamoV1

Data marker for collation jamo data.

CollationMetadata

Each non-alias collation that the data provider knows about explicitly has an data entry at least for this struct.

CollationMetadataV1

Data marker for collation metadata.

CollationReordering

Script reordering data

CollationReorderingV1

Data marker for collation reordering data.

CollationRootV1

Data marker for singleton root collation data.

CollationSpecialPrimaries

Special primaries associated with the root collation

CollationSpecialPrimariesV1

Data marker for collcation special primaries data.

CollationTailoringV1

Data marker for collation tailorings.