icu::collator

Struct Collator

pub struct Collator { /* private fields */ }

Compares strings according to culturally-relevant ordering.

Implementations

impl Collator

pub fn as_borrowed(&self) -> CollatorBorrowed<'_>

Constructs a borrowed version of this type for more efficient querying.

pub fn try_new( prefs: CollatorPreferences, options: CollatorOptions, ) -> Result<CollatorBorrowed<'static>, DataError>

Creates CollatorBorrowed for the given locale and options from compiled data.

pub fn try_new_with_buffer_provider( provider: &(impl BufferProvider + ?Sized), prefs: CollatorPreferences, options: CollatorOptions, ) -> Result<Collator, DataError>

A version of [Self :: try_new] that uses custom data provided by a BufferProvider.

โจ Enabled with the serde feature.

๐ Help choosing a constructor

pub fn try_new_unstable<D>( provider: &D, prefs: CollatorPreferences, options: CollatorOptions, ) -> Result<Collator, DataError>
where D: DataProvider<CollationSpecialPrimariesV1> + DataProvider<CollationRootV1> + DataProvider<CollationTailoringV1> + DataProvider<CollationDiacriticsV1> + DataProvider<CollationJamoV1> + DataProvider<CollationMetadataV1> + DataProvider<CollationReorderingV1> + DataProvider<NormalizerNfdDataV1> + DataProvider<NormalizerNfdTablesV1> + ?Sized,

A version of Self::try_new that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

โ๏ธ The bounds on provider may change over time, including in SemVer minor releases.