Struct icu::collections::codepointinvlist::CodePointInversionList

pub struct CodePointInversionList<'data> { /* private fields */ }

A membership wrapper for CodePointInversionList.

Provides exposure to membership functions and constructors from serialized CodePointSets (sets of code points) and predefined ranges.

Implementations

impl<'data> CodePointInversionList<'data>

pub fn try_from_inversion_list( inv_list: ZeroVec<'data, u32> ) -> Result<CodePointInversionList<'data>, CodePointInversionListError>

Returns a new CodePointInversionList from an inversion list represented as a ZeroVec<u32> of code points.

The inversion list must be of even length, sorted ascending non-overlapping, and within the bounds of 0x0 -> 0x10FFFF inclusive, and end points being exclusive.

Examples

use icu_collections::codepointinvlist::CodePointInversionList;
use icu_collections::codepointinvlist::CodePointInversionListError;
use zerovec::ZeroVec;
let valid = [0x0, 0x10000];
let inv_list: ZeroVec<u32> = ZeroVec::from_slice_or_alloc(&valid);
let result = CodePointInversionList::try_from_inversion_list(inv_list);
assert!(matches!(result, CodePointInversionList));

let invalid: Vec<u32> = vec![0x0, 0x80, 0x3];
let inv_list: ZeroVec<u32> = ZeroVec::from_slice_or_alloc(&invalid);
let result = CodePointInversionList::try_from_inversion_list(inv_list);
assert!(matches!(
    result,
    Err(CodePointInversionListError::InvalidSet(_))
));
if let Err(CodePointInversionListError::InvalidSet(actual)) = result {
    assert_eq!(&invalid, &actual);
}

pub fn try_from_inversion_list_slice( inv_list: &'data [u32] ) -> Result<CodePointInversionList<'data>, CodePointInversionListError>

Returns a new CodePointInversionList by borrowing an inversion list represented as a slice of u32 code points.

The inversion list must be of even length, sorted ascending non-overlapping, and within the bounds of 0x0 -> 0x10FFFF inclusive, and end points being exclusive.

Note: The slice may be cloned on certain platforms; for more information, see ZeroVec::from_slice_or_alloc.

Examples

use icu_collections::codepointinvlist::CodePointInversionList;
use icu_collections::codepointinvlist::CodePointInversionListError;
use zerovec::ZeroVec;
let valid = [0x0, 0x10000];
let result = CodePointInversionList::try_from_inversion_list_slice(&valid);
assert!(matches!(result, CodePointInversionList));

let invalid: Vec<u32> = vec![0x0, 0x80, 0x3];
let result =
    CodePointInversionList::try_from_inversion_list_slice(&invalid);
assert!(matches!(
    result,
    Err(CodePointInversionListError::InvalidSet(_))
));
if let Err(CodePointInversionListError::InvalidSet(actual)) = result {
    assert_eq!(&invalid, &actual);
}

pub fn try_clone_from_inversion_list_slice( inv_list: &[u32] ) -> Result<CodePointInversionList<'data>, CodePointInversionListError>

Returns a new, fully-owned CodePointInversionList by cloning an inversion list represented as a slice of u32 code points.

The inversion list must be of even length, sorted ascending non-overlapping, and within the bounds of 0x0 -> 0x10FFFF inclusive, and end points being exclusive.

Examples

use icu_collections::codepointinvlist::CodePointInversionList;
use std::vec::Vec;
use zerovec::ZeroVec;

let bmp_list = &[0x0, 0x10000];
let smp_list = &[0x10000, 0x20000];
let sip_list = &[0x20000, 0x30000];

let lists: Vec<CodePointInversionList> =
    [&bmp_list[..], smp_list, sip_list]
        .into_iter()
        .map(|l| {
            CodePointInversionList::try_clone_from_inversion_list_slice(l)
                .unwrap()
        })
        .collect();

let bmp = &lists[0];
assert!(bmp.contains32(0xFFFF));
assert!(!bmp.contains32(0x10000));

assert!(!lists.iter().any(|set| set.contains32(0x40000)));

pub fn get_inversion_list_vec(&self) -> Vec<u32, Global>

Returns an owned inversion list representing the current CodePointInversionList

pub fn all() -> CodePointInversionList<'data>

Returns CodePointInversionList spanning entire Unicode range

The range spans from 0x0 -> 0x10FFFF inclusive.

Examples

use icu_collections::codepointinvlist::CodePointInversionList;
use zerovec::ZeroVec;

let expected = vec![0x0, (char::MAX as u32) + 1];
assert_eq!(
    CodePointInversionList::all().get_inversion_list_vec(),
    expected
);
assert_eq!(
    CodePointInversionList::all().size(),
    (expected[1] - expected[0]) as usize
);

pub fn bmp() -> CodePointInversionList<'data>

Returns CodePointInversionList spanning BMP range

The range spans from 0x0 -> 0xFFFF inclusive.

Examples

use icu_collections::codepointinvlist::CodePointInversionList;
use zerovec::ZeroVec;

const BMP_MAX: u32 = 0xFFFF;

let expected = vec![0x0, BMP_MAX + 1];
assert_eq!(
    CodePointInversionList::bmp().get_inversion_list_vec(),
    expected
);
assert_eq!(
    CodePointInversionList::bmp().size(),
    (expected[1] - expected[0]) as usize
);

pub fn iter_chars(&self) -> impl Iterator<Item = char>

Yields an Iterator going through the character set in the CodePointInversionList

Examples

use icu_collections::codepointinvlist::CodePointInversionList;
let example_list = [0x41, 0x44, 0x45, 0x46];
let example =
    CodePointInversionList::try_from_inversion_list_slice(&example_list)
        .unwrap();
let mut ex_iter_chars = example.iter_chars();
assert_eq!(Some('A'), ex_iter_chars.next());
assert_eq!(Some('B'), ex_iter_chars.next());
assert_eq!(Some('C'), ex_iter_chars.next());
assert_eq!(Some('E'), ex_iter_chars.next());
assert_eq!(None, ex_iter_chars.next());

pub fn iter_ranges(&self) -> impl ExactSizeIterator

Yields an Iterator returning the ranges of the code points that are included in the CodePointInversionList

Ranges are returned as RangeInclusive, which is inclusive of its end bound value. An end-inclusive behavior matches the ICU4C/J behavior of ranges, ex: CodePointInversionList::contains(UChar32 start, UChar32 end).

Example

use icu_collections::codepointinvlist::CodePointInversionList;
let example_list = [0x41, 0x44, 0x45, 0x46];
let example =
    CodePointInversionList::try_from_inversion_list_slice(&example_list)
        .unwrap();
let mut example_iter_ranges = example.iter_ranges();
assert_eq!(Some(0x41..=0x43), example_iter_ranges.next());
assert_eq!(Some(0x45..=0x45), example_iter_ranges.next());
assert_eq!(None, example_iter_ranges.next());

pub fn iter_ranges_complemented( &self ) -> impl Iterator<Item = RangeInclusive<u32>>

Yields an Iterator returning the ranges of the code points that are not included in the CodePointInversionList

Ranges are returned as RangeInclusive, which is inclusive of its end bound value. An end-inclusive behavior matches the ICU4C/J behavior of ranges, ex: CodePointInversionList::contains(UChar32 start, UChar32 end).

Example

use icu_collections::codepointinvlist::CodePointInversionList;
let example_list = [0x41, 0x44, 0x45, 0x46];
let example =
    CodePointInversionList::try_from_inversion_list_slice(&example_list)
        .unwrap();
let mut example_iter_ranges = example.iter_ranges_complemented();
assert_eq!(Some(0..=0x40), example_iter_ranges.next());
assert_eq!(Some(0x44..=0x44), example_iter_ranges.next());
assert_eq!(Some(0x46..=char::MAX as u32), example_iter_ranges.next());
assert_eq!(None, example_iter_ranges.next());

pub fn get_range_count(&self) -> usize

Returns the number of ranges contained in this CodePointInversionList

pub fn get_nth_range(&self, idx: usize) -> Option<RangeInclusive<u32>>

Returns a specific range contained in this CodePointInversionList by index. Intended for use in FFI.

pub fn size(&self) -> usize

Returns the number of elements of the CodePointInversionList

pub fn is_empty(&self) -> bool

Returns whether or not the CodePointInversionList is empty

pub fn contains(&self, query: char) -> bool

Checks to see the query is in the CodePointInversionList

Runs a binary search in O(log(n)) where n is the number of start and end points in the set using core implementation

Examples

use icu_collections::codepointinvlist::CodePointInversionList;
let example_list = [0x41, 0x43, 0x44, 0x45];
let example =
    CodePointInversionList::try_from_inversion_list_slice(&example_list)
        .unwrap();
assert!(example.contains('A'));
assert!(!example.contains('C'));

pub fn contains32(&self, query: u32) -> bool

Checks to see the unsigned int is in the CodePointInversionList::all()

Note: Even though u32 and char in Rust are non-negative 4-byte values, there is an important difference. A u32 can take values up to a very large integer value, while a char in Rust is defined to be in the range from 0 to the maximum valid Unicode Scalar Value.

Runs a binary search in O(log(n)) where n is the number of start and end points in the set using core implementation

Examples

use icu_collections::codepointinvlist::CodePointInversionList;
let example_list = [0x41, 0x43, 0x44, 0x45];
let example =
    CodePointInversionList::try_from_inversion_list_slice(&example_list)
        .unwrap();
assert!(example.contains32(0x41));
assert!(!example.contains32(0x43));

pub fn contains_range(&self, range: &impl RangeBounds<char>) -> bool

Checks to see if the range is in the CodePointInversionList

Runs a binary search in O(log(n)) where n is the number of start and end points in the set using Vec implementation. Only runs the search once on the start parameter, while the end parameter is checked in a single O(1) step.

Examples

use icu_collections::codepointinvlist::CodePointInversionList;
let example_list = [0x41, 0x43, 0x44, 0x45];
let example =
    CodePointInversionList::try_from_inversion_list_slice(&example_list)
        .unwrap();
assert!(example.contains_range(&('A'..'C')));
assert!(example.contains_range(&('A'..='B')));
assert!(!example.contains_range(&('A'..='C')));

Surrogate points (0xD800 -> 0xDFFF) will return false if the Range contains them but the CodePointInversionList does not.

Note: when comparing to ICU4C/J, keep in mind that Ranges in Rust are constructed inclusive of start boundary and exclusive of end boundary. The ICU4C/J CodePointInversionList::contains(UChar32 start, UChar32 end) method differs by including the end boundary.

Examples

use icu_collections::codepointinvlist::CodePointInversionList;
use std::char;
let check =
    char::from_u32(0xD7FE).unwrap()..char::from_u32(0xE001).unwrap();
let example_list = [0xD7FE, 0xD7FF, 0xE000, 0xE001];
let example =
    CodePointInversionList::try_from_inversion_list_slice(&example_list)
        .unwrap();
assert!(!example.contains_range(&(check)));

pub fn contains_set(&self, set: &CodePointInversionList<'data>) -> bool

Check if the calling CodePointInversionList contains all the characters of the given CodePointInversionList

Examples

use icu_collections::codepointinvlist::CodePointInversionList;
let example_list = [0x41, 0x46, 0x55, 0x5B]; // A - E, U - Z
let example =
    CodePointInversionList::try_from_inversion_list_slice(&example_list)
        .unwrap();
let a_to_d =
    CodePointInversionList::try_from_inversion_list_slice(&[0x41, 0x45])
        .unwrap();
let f_to_t =
    CodePointInversionList::try_from_inversion_list_slice(&[0x46, 0x55])
        .unwrap();
let r_to_x =
    CodePointInversionList::try_from_inversion_list_slice(&[0x52, 0x58])
        .unwrap();
assert!(example.contains_set(&a_to_d)); // contains all
assert!(!example.contains_set(&f_to_t)); // contains none
assert!(!example.contains_set(&r_to_x)); // contains some

pub fn span(&self, span_str: &str, contained: bool) -> usize

Returns the end of the initial substring where the characters are either contained/not contained in the set.

Examples

use icu_collections::codepointinvlist::CodePointInversionList;
let example_list = [0x41, 0x44]; // {A, B, C}
let example =
    CodePointInversionList::try_from_inversion_list_slice(&example_list)
        .unwrap();
assert_eq!(example.span("CABXYZ", true), 3);
assert_eq!(example.span("XYZC", false), 3);
assert_eq!(example.span("XYZ", true), 0);
assert_eq!(example.span("ABC", false), 0);

pub fn span_back(&self, span_str: &str, contained: bool) -> usize

Returns the start of the trailing substring (starting from end of string) where the characters are either contained/not contained in the set. Returns the length of the string if no valid return.

Examples

use icu_collections::codepointinvlist::CodePointInversionList;
let example_list = [0x41, 0x44]; // {A, B, C}
let example =
    CodePointInversionList::try_from_inversion_list_slice(&example_list)
        .unwrap();
assert_eq!(example.span_back("XYZCAB", true), 3);
assert_eq!(example.span_back("ABCXYZ", true), 6);
assert_eq!(example.span_back("CABXYZ", false), 3);