Trait icu::datetime::date::DateTimeType

pub trait DateTimeType: FromStr {
    fn year(&self) -> usize;

    fn month(&self) -> Month;

    fn day(&self) -> Day;

    fn hour(&self) -> Hour;

    fn minute(&self) -> Minute;

    fn second(&self) -> Second;
}

Temporary trait used to represent the input data for DateTimeFormat.

This type represents all data that the formatted needs in order to produced formatted string.

Note: At the moment we support only gregorian calendar, and plan to extend support to other calendars in the upcoming releases. See https://github.com/unicode-org/icu4x/issues/355

Required methods

fn year(&self) -> usize

fn month(&self) -> Month

fn day(&self) -> Day

fn hour(&self) -> Hour

fn minute(&self) -> Minute

fn second(&self) -> Second

Loading content...