Enum icu::datetime::DateTimeFormatError

pub enum DateTimeFormatError {
    Pattern(PatternError),
    Format(Error),
    DataProvider(DataError),
    MissingInputField,
    Skeleton(SkeletonError),
    UnsupportedField(FieldSymbol),
    UnsupportedOptions,
    PluralRules(PluralRulesError),
    DateTimeInput(DateTimeError),
    MissingWeekdaySymbol(usize),
    MissingMonthSymbol(usize),
    MissingEraSymbol(TinyAsciiStr<16_usize>),
}

A list of possible error outcomes for the DateTimeFormat struct.

Variants

Pattern(PatternError)

Tuple Fields

0: PatternError

An error originating from parsing a pattern.

Format(Error)

Tuple Fields

0: Error

An error originating from the Write trait.

DataProvider(DataError)

Tuple Fields

0: DataError

An error originating inside of the data provider.

MissingInputField

An error originating from a missing field in datetime input. TODO: How can we return which field was missing?

Skeleton(SkeletonError)

Tuple Fields

0: SkeletonError

An error originating from skeleton matching.

UnsupportedField(FieldSymbol)

Tuple Fields

0: FieldSymbol

An error originating from an unsupported field in a datetime format.

UnsupportedOptions

An error due to there being no patterns for the given options.

PluralRules(PluralRulesError)

Tuple Fields

0: PluralRulesError

An error originating from PluralRules.

DateTimeInput(DateTimeError)

Tuple Fields

0: DateTimeError

An error originating from DateTimeInput.

MissingWeekdaySymbol(usize)

Tuple Fields

0: usize

An error originating from a missing weekday symbol in the data.

MissingMonthSymbol(usize)

Tuple Fields

0: usize

An error originating from a missing month symbol in the data.

MissingEraSymbol(TinyAsciiStr<16_usize>)

Tuple Fields

0: TinyAsciiStr<16_usize>

An error originating from a missing era symbol in the data.