Enum icu::datetime::DateTimeFormatterOptions

pub enum DateTimeFormatterOptions {
    Length(Bag),
}

A bag of options which, together with Locale, defines how dates will be formatted with a TypedDateTimeFormatter instance.

Each variant of the bag is a combination of settings defining how to format the date, with an optional Preferences which represent user preferences and may alter how the selected pattern is formatted.

Examples

use icu::datetime::{options::length, DateTimeFormatterOptions};

let bag = length::Bag::from_date_time_style(
    length::Date::Medium,
    length::Time::Short,
);

At the moment only the length::Bag works, and we plan to extend that to support ECMA402 like components bag later.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Length(Bag)

Bag of lengths for date and time.