Enum icu::datetime::fields::Second

#[repr(u8)]
pub enum Second {
    Second,
    FractionalSecond,
    Millisecond,
}

An enum for the possible symbols of a second field in a date pattern.

Variants

Second

Field symbol for second (numeric).

This field symbol is represented by the character s in a date formatting pattern string.

For more details, see documentation on date field symbols.

FractionalSecond

Field symbol for fractional second (numeric).

Produces the number of digits specified by the field length.

This field symbol is represented by the character S in a date formatting pattern string.

For more details, see documentation on date field symbols.

Millisecond

Field symbol for milliseconds in day (numeric).

This field behaves exactly like a composite of all time-related fields, not including the zone fields.

This field symbol is represented by the character A in a date formatting pattern string.

For more details, see documentation on date field symbols.