Enum icu::datetime::fields::Week

#[repr(u8)]
pub enum Week {
    WeekOfYear = 0,
    WeekOfMonth = 1,
}

An enum for the possible symbols of a week field in a date pattern.

Variants

WeekOfYear = 0

Field symbol for week of year (numeric).

When used in a pattern with year, use Year::WeekOf for the year field instead of Year::Calendar.

This field symbol is represented by the character w in a date formatting pattern string.

For more details, see documentation on date field symbols.

WeekOfMonth = 1

Field symbol for week of month (numeric).

This field symbol is represented by the character W in a date formatting pattern string.

For more details, see documentation on date field symbols.