icu::datetime::fieldsets

Struct Combo

pub struct Combo<DT, Z> { /* private fields */ }

Struct for combining date/time fields with zone fields.

This struct produces “composite field sets” as defined in UTS 35. See fieldsets.

Examples

The only way to construct a combo field set (in this case, weekday with location-based zone):

use icu::datetime::fieldsets::{zone::Location, Combo, E};

let field_set = E::long().with_zone(Location);

Format the weekday, hour, and location-based zone:

use icu::datetime::fieldsets::{self, zone, Combo};
use icu::datetime::input::ZonedDateTime;
use icu::datetime::DateTimeFormatter;
use icu::locale::locale;
use icu::time::zone::IanaParser;
use writeable::assert_writeable_eq;

// Note: Combo type can be elided, but it is shown here for demonstration
let formatter =
    DateTimeFormatter::<Combo<fieldsets::ET, zone::Location>>::try_new(
        locale!("en-US").into(),
        fieldsets::E::short()
            .with_time_hm()
            .with_zone(zone::Location),
    )
    .unwrap();

let zdt = ZonedDateTime::try_location_only_from_str(
    "2024-10-18T15:44[America/Los_Angeles]",
    formatter.calendar(),
    IanaParser::new(),
)
.unwrap();

assert_writeable_eq!(
    formatter.format(&zdt),
    "Fri, 3:44 PM Los Angeles Time"
);

Same thing with a fixed calendar formatter:

use icu::calendar::Gregorian;
use icu::datetime::fieldsets::{self, zone, Combo};
use icu::datetime::input::ZonedDateTime;
use icu::datetime::FixedCalendarDateTimeFormatter;
use icu::locale::locale;
use icu::time::zone::IanaParser;
use writeable::assert_writeable_eq;

// Note: Combo type can be elided, but it is shown here for demonstration
let formatter = FixedCalendarDateTimeFormatter::<
    _,
    Combo<fieldsets::ET, zone::Location>,
>::try_new(
    locale!("en-US").into(),
    fieldsets::E::short()
        .with_time_hm()
        .with_zone(zone::Location),
)
.unwrap();

let zdt = ZonedDateTime::try_location_only_from_str(
    "2024-10-18T15:44[America/Los_Angeles]",
    Gregorian,
    IanaParser::new(),
)
.unwrap();

assert_writeable_eq!(
    formatter.format(&zdt),
    "Fri, 3:44 PM Los Angeles Time"
);

Mix a dynamic DateFieldSet with a static time zone:

use icu::datetime::fieldsets::{
    enums::DateFieldSet, zone::GenericShort, Combo, YMD,
};
use icu::datetime::input::ZonedDateTime;
use icu::datetime::DateTimeFormatter;
use icu::locale::locale;
use icu::time::zone::IanaParser;
use writeable::assert_writeable_eq;

// Note: Combo type can be elided, but it is shown here for demonstration
let formatter =
    DateTimeFormatter::<Combo<DateFieldSet, GenericShort>>::try_new(
        locale!("en-US").into(),
        DateFieldSet::YMD(YMD::long()).with_zone(GenericShort),
    )
    .unwrap();

let zdt = ZonedDateTime::try_location_only_from_str(
    "2024-10-18T15:44[America/Los_Angeles]",
    formatter.calendar(),
    IanaParser::new(),
)
.unwrap();

assert_writeable_eq!(formatter.format(&zdt), "October 18, 2024 PT");

Format with a time of day and long time zone:

use icu::calendar::Gregorian;
use icu::datetime::fieldsets::{zone::SpecificLong, T};
use icu::datetime::input::ZonedDateTime;
use icu::datetime::FixedCalendarDateTimeFormatter;
use icu::locale::locale;
use icu::time::zone::IanaParser;
use writeable::assert_writeable_eq;

let formatter = FixedCalendarDateTimeFormatter::try_new(
    locale!("en-US").into(),
    T::medium().with_zone(SpecificLong),
)
.unwrap();

let zdt = ZonedDateTime::try_strict_from_str(
    "2024-10-18T15:44-0700[America/Los_Angeles]",
    Gregorian,
    IanaParser::new(),
)
.unwrap();

assert_writeable_eq!(
    formatter.format(&zdt),
    "3:44:00 PM Pacific Daylight Time"
);

Implementations

impl Combo<D, SpecificLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<D, SpecificShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<D, LocalizedOffsetLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<D, LocalizedOffsetShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<D, GenericLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<D, GenericShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<D, Location>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<D, ExemplarCity>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<D, ZoneFieldSet>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DT, SpecificLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DT, SpecificShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DT, LocalizedOffsetLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DT, LocalizedOffsetShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DT, GenericLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DT, GenericShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DT, Location>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DT, ExemplarCity>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DT, ZoneFieldSet>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<E, SpecificLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<E, SpecificShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<E, LocalizedOffsetLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<E, LocalizedOffsetShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<E, GenericLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<E, GenericShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<E, Location>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<E, ExemplarCity>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<E, ZoneFieldSet>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<ET, SpecificLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<ET, SpecificShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<ET, LocalizedOffsetLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<ET, LocalizedOffsetShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<ET, GenericLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<ET, GenericShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<ET, Location>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<ET, ExemplarCity>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<ET, ZoneFieldSet>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DE, SpecificLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DE, SpecificShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DE, LocalizedOffsetLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DE, LocalizedOffsetShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DE, GenericLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DE, GenericShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DE, Location>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DE, ExemplarCity>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DE, ZoneFieldSet>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DET, SpecificLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DET, SpecificShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DET, LocalizedOffsetLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DET, LocalizedOffsetShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DET, GenericLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DET, GenericShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DET, Location>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DET, ExemplarCity>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DET, ZoneFieldSet>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MD, SpecificLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MD, SpecificShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MD, LocalizedOffsetLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MD, LocalizedOffsetShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MD, GenericLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MD, GenericShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MD, Location>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MD, ExemplarCity>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MD, ZoneFieldSet>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDT, SpecificLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDT, SpecificShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDT, LocalizedOffsetLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDT, LocalizedOffsetShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDT, GenericLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDT, GenericShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDT, Location>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDT, ExemplarCity>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDT, ZoneFieldSet>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDE, SpecificLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDE, SpecificShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDE, LocalizedOffsetLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDE, LocalizedOffsetShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDE, GenericLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDE, GenericShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDE, Location>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDE, ExemplarCity>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDE, ZoneFieldSet>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDET, SpecificLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDET, SpecificShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDET, LocalizedOffsetLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDET, LocalizedOffsetShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDET, GenericLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDET, GenericShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDET, Location>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDET, ExemplarCity>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<MDET, ZoneFieldSet>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMD, SpecificLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMD, SpecificShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMD, LocalizedOffsetLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMD, LocalizedOffsetShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMD, GenericLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMD, GenericShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMD, Location>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMD, ExemplarCity>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMD, ZoneFieldSet>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDT, SpecificLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDT, SpecificShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDT, LocalizedOffsetLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDT, LocalizedOffsetShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDT, GenericLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDT, GenericShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDT, Location>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDT, ExemplarCity>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDT, ZoneFieldSet>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDE, SpecificLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDE, SpecificShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDE, LocalizedOffsetLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDE, LocalizedOffsetShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDE, GenericLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDE, GenericShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDE, Location>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDE, ExemplarCity>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDE, ZoneFieldSet>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDET, SpecificLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDET, SpecificShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDET, LocalizedOffsetLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDET, LocalizedOffsetShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDET, GenericLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDET, GenericShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDET, Location>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDET, ExemplarCity>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<YMDET, ZoneFieldSet>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<T, SpecificLong>

pub fn into_enums(self) -> Combo<TimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<T, SpecificShort>

pub fn into_enums(self) -> Combo<TimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<T, LocalizedOffsetLong>

pub fn into_enums(self) -> Combo<TimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<T, LocalizedOffsetShort>

pub fn into_enums(self) -> Combo<TimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<T, GenericLong>

pub fn into_enums(self) -> Combo<TimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<T, GenericShort>

pub fn into_enums(self) -> Combo<TimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<T, Location>

pub fn into_enums(self) -> Combo<TimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<T, ExemplarCity>

pub fn into_enums(self) -> Combo<TimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<T, ZoneFieldSet>

pub fn into_enums(self) -> Combo<TimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DateFieldSet, SpecificLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DateFieldSet, SpecificShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DateFieldSet, LocalizedOffsetLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DateFieldSet, LocalizedOffsetShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DateFieldSet, GenericLong>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DateFieldSet, GenericShort>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DateFieldSet, Location>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DateFieldSet, ExemplarCity>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DateFieldSet, ZoneFieldSet>

pub fn into_enums(self) -> Combo<DateFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<TimeFieldSet, SpecificLong>

pub fn into_enums(self) -> Combo<TimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<TimeFieldSet, SpecificShort>

pub fn into_enums(self) -> Combo<TimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<TimeFieldSet, LocalizedOffsetLong>

pub fn into_enums(self) -> Combo<TimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<TimeFieldSet, LocalizedOffsetShort>

pub fn into_enums(self) -> Combo<TimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<TimeFieldSet, GenericLong>

pub fn into_enums(self) -> Combo<TimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<TimeFieldSet, GenericShort>

pub fn into_enums(self) -> Combo<TimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<TimeFieldSet, Location>

pub fn into_enums(self) -> Combo<TimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<TimeFieldSet, ExemplarCity>

pub fn into_enums(self) -> Combo<TimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<TimeFieldSet, ZoneFieldSet>

pub fn into_enums(self) -> Combo<TimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DateAndTimeFieldSet, SpecificLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DateAndTimeFieldSet, SpecificShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DateAndTimeFieldSet, LocalizedOffsetLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DateAndTimeFieldSet, LocalizedOffsetShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DateAndTimeFieldSet, GenericLong>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DateAndTimeFieldSet, GenericShort>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DateAndTimeFieldSet, Location>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DateAndTimeFieldSet, ExemplarCity>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.

impl Combo<DateAndTimeFieldSet, ZoneFieldSet>

pub fn into_enums(self) -> Combo<DateAndTimeFieldSet, ZoneFieldSet>

Convert this specific Combo into a more general Combo. Useful when adding to the field of a CompositeFieldSet.