icu::datetime::fieldsets

Struct YMDET

#[non_exhaustive]
pub struct YMDET {
    pub length: Length,
    pub alignment: Option<Alignment>,
    pub year_style: Option<YearStyle>,
    pub time_precision: Option<TimePrecision>,
}

“Fri, 5/17/24, 3:47:50 PM” ⇒ year, month, day, and weekday with time

Examples

In DateTimeFormatter:

use icu::datetime::input::Date;
use icu::datetime::DateTimeFormatter;
use icu::datetime::fieldsets::YMDET;
use icu::locale::locale;
use icu::datetime::input::{DateTime, Time};
use writeable::assert_writeable_eq;

let fmt = DateTimeFormatter::try_new(
    locale!("en").into(),
   YMDET::short(),
)
.unwrap();
let dt = DateTime { date: Date::try_new_iso(2024, 5, 17).unwrap(), time: Time::try_new(15, 47, 50, 0).unwrap() };

assert_writeable_eq!(
    fmt.format(&dt),
   "Fri, 5/17/24, 3:47:50 PM"
);

In FixedCalendarDateTimeFormatter:

use icu::datetime::input::Date;
use icu::calendar::Gregorian;
use icu::datetime::FixedCalendarDateTimeFormatter;
use icu::datetime::fieldsets::YMDET;
use icu::locale::locale;
use icu::datetime::input::{DateTime, Time};
use writeable::assert_writeable_eq;

let fmt = FixedCalendarDateTimeFormatter::try_new(
    locale!("en").into(),
   YMDET::short(),
)
.unwrap();
let dt = DateTime { date: Date::try_new_gregorian(2024, 5, 17).unwrap(), time: Time::try_new(15, 47, 50, 0).unwrap() };

assert_writeable_eq!(
    fmt.format(&dt),
   "Fri, 5/17/24, 3:47:50 PM"
);

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

length: Length

The desired length of the formatted string.

See: Length

alignment: Option<Alignment>

Whether fields should be aligned for a column-like layout.

See: Alignment

year_style: Option<YearStyle>

When to display the era field in the formatted string.

See: YearStyle

time_precision: Option<TimePrecision>

How precisely to display the time of day

See: TimePrecision

Implementations

impl YMDET

pub const fn with_length(self, length: Length) -> YMDET

Sets the length option.

impl YMDET

pub const fn with_alignment(self, alignment: Alignment) -> YMDET

Sets the alignment option.

impl YMDET

pub const fn with_year_style(self, year_style: YearStyle) -> YMDET

Sets the year style option.

impl YMDET

pub const fn with_time_precision(self, time_precision: TimePrecision) -> YMDET

Sets the time precision option.

impl YMDET

pub const fn for_length(length: Length) -> YMDET

Creates a YMDET skeleton with the given formatting length.

pub const fn long() -> YMDET

Creates a YMDET skeleton with a long length.

pub const fn medium() -> YMDET

Creates a YMDET skeleton with a medium length.

pub const fn short() -> YMDET

Creates a YMDET skeleton with a short length.

impl YMDET

pub fn with_zone<Z>(self, zone: Z) -> Combo<YMDET, Z>
where Z: ZoneMarkers,

Associates this field set with a time zone field set.