icu::datetime::input

Struct Time

pub struct Time {
    pub hour: Hour,
    pub minute: Minute,
    pub second: Second,
    pub subsecond: Nanosecond,
}

A representation of a time in hours, minutes, seconds, and nanoseconds

The primary definition of this type is in the icu_time crate. Other ICU4X crates re-export it for convenience.

This type supports the range [00:00:00.000000000, 23:59:60.999999999].

Fields
hour: Hour

Hour

minute: Minute

Minute

second: Second

Second

subsecond: Nanosecond

Subsecond

Implementations

impl Time

pub fn try_from_str(rfc_9557_str: &str) -> Result<Time, ParseError>

Creates a Time from an RFC 9557 string of a time.

Does not support parsing an RFC 9557 string with a date and time; for that, use DateTime.

✨ Enabled with the ixdtf Cargo feature.

Examples

use icu::time::Time;

let time = Time::try_from_str("16:01:17.045").unwrap();

assert_eq!(time.hour.number(), 16);
assert_eq!(time.minute.number(), 1);
assert_eq!(time.second.number(), 17);
assert_eq!(time.subsecond.number(), 45000000);

pub fn try_from_utf8(rfc_9557_str: &[u8]) -> Result<Time, ParseError>

Creates a Time in the ISO-8601 calendar from an RFC 9557 string.

✨ Enabled with the ixdtf Cargo feature.

See Self::try_from_str().

impl Time

pub const fn new( hour: Hour, minute: Minute, second: Second, subsecond: Nanosecond, ) -> Time

Construct a new Time, without validating that all components are in range

pub const fn start_of_day() -> Time

Construct a new Time representing the start of the day (00:00:00.000)

pub const fn noon() -> Time

Construct a new Time representing noon (12:00:00.000)

pub fn try_new( hour: u8, minute: u8, second: u8, nanosecond: u32, ) -> Result<Time, RangeError>

Construct a new Time, whilst validating that all components are in range