Module icu::datetime::options::preferences

Preferences is a bag of options to be associated with either Length or Components bag which provides information on user preferences that can affect the result of the formatting.

Unicode Extensions

User preferences will often be stored as part of the Locale identified as Unicode Extensions, but for scenarios where the application stores information about user preferences they can be also provided via this bag (and if they are, they will take precedence over unicode extensions from the locale).

Examples

use icu_datetime::options::preferences;

let prefs = preferences::Bag {
    hour_cycle: Some(preferences::HourCycle::H23)
};

Structs

Bag
Bag of preferences stores user preferences which may affect the result of date and time formatting.

Enums

HourCycle
User Preference for adjusting how hour component is displayed.