icu::datetime::pattern

Struct DateTimePattern

pub struct DateTimePattern { /* private fields */ }

A pattern for formatting a datetime in a calendar.

โ This type forgoes most internationalization functionality of the datetime crate. It assumes that the pattern is already localized for the customerโs locale. Most clients should use DateTimeFormatter instead of directly formatting with patterns.

There are two ways to make one of these:

From a custom pattern string: DateTimePattern::try_from_pattern_str

From a formatted datetime: FormattedDateTime::pattern

Things you can do with one of these:

Use it to directly format a datetime via FixedCalendarDateTimeNames

Convert it to a string pattern via Writeable

Get the resolved components

๐ This item has a stack size of 32 bytes on the stable toolchain at release date.

Examples

Create a pattern from a custom string and compare it to one from data, then check the resolved components:

use icu::calendar::Gregorian;
use icu::datetime::fieldsets::YMD;
use icu::datetime::input::Date;
use icu::datetime::pattern::DateTimePattern;
use icu::datetime::provider::fields::components;
use icu::datetime::FixedCalendarDateTimeFormatter;
use icu::locale::locale;
use writeable::assert_writeable_eq;

// Create the pattern from a string:
let pattern_str = "d MMM y";
let custom_pattern =
    DateTimePattern::try_from_pattern_str(pattern_str).unwrap();
assert_writeable_eq!(custom_pattern, pattern_str);

// Load data that resolves to the same pattern:
let data_pattern = FixedCalendarDateTimeFormatter::<Gregorian, _>::try_new(
    locale!("es-MX").into(),
    YMD::medium(),
)
.unwrap()
// The pattern can depend on the datetime being formatted.
.format(&Date::try_new_gregorian(2024, 1, 1).unwrap())
.pattern();
assert_writeable_eq!(data_pattern, pattern_str);
assert_eq!(custom_pattern, data_pattern);

// Check the resolved components:
let mut expected_components_bag = components::Bag::default();
expected_components_bag.year = Some(components::Year::Numeric);
expected_components_bag.month = Some(components::Month::Short);
expected_components_bag.day = Some(components::Day::NumericDayOfMonth);
let actual_components_bag = components::Bag::from(&data_pattern);
assert_eq!(actual_components_bag, expected_components_bag);

Check the hour cycle of a resolved pattern:

use icu::datetime::fieldsets::T;
use icu::datetime::input::Time;
use icu::datetime::pattern::DateTimePattern;
use icu::datetime::provider::fields::components;
use icu::datetime::NoCalendarFormatter;
use icu::locale::locale;
use icu::locale::preferences::extensions::unicode::keywords::HourCycle;
use writeable::assert_writeable_eq;

let pattern =
    NoCalendarFormatter::try_new(locale!("es-MX").into(), T::medium())
        .unwrap()
        // The pattern can depend on the datetime being formatted.
        .format(&Time::try_new(12, 0, 0, 0).unwrap())
        .pattern();

assert_writeable_eq!(pattern, "hh:mm:ssโฏa");

// Get the hour cycle from the resolved components:
let components = components::Bag::from(&pattern);
assert_eq!(components.hour_cycle, Some(HourCycle::H12));

Implementations

impl DateTimePattern

pub fn try_from_pattern_str( pattern_str: &str, ) -> Result<DateTimePattern, PatternError>

Creates a DateTimePattern from a pattern string.

For more details on the syntax, see UTS 35: https://unicode.org/reports/tr35/tr35-dates.html#Date_Format_Patterns

impl DateTimePattern

pub fn to_string(&self) -> String

Converts the given value to a String.

Under the hood, this uses an efficient Writeable implementation. However, in order to avoid allocating a string, it is more efficient to use Writeable directly.