Module icu::datetime::provider::calendar

Data structs for calendar-specific symbols and patterns.

Modules

day_periods

Formatting symbols for DayPeriod.

months

Formatting symbols for Month.

patterns

Data structs for date / time patterns that store data corresponding to pattern lengths and/or plural forms.

weekdays

Formatting symbols for Weekday.

Structs

BuddhistDateLengthsV1Marker

Marker type for DateLengthsV1: “datetime/buddhist/datelengths@1”

BuddhistDateSymbolsV1Marker

Marker type for DateSymbolsV1: “datetime/buddhist/datesymbols@1”

ChineseDateLengthsV1Marker

Marker type for DateLengthsV1: “datetime/chinese/datelengths@1”

ChineseDateSymbolsV1Marker

Marker type for DateSymbolsV1: “datetime/chinese/datesymbols@1”

CopticDateLengthsV1Marker

Marker type for DateLengthsV1: “datetime/coptic/datelengths@1”

CopticDateSymbolsV1Marker

Marker type for DateSymbolsV1: “datetime/coptic/datesymbols@1”

DangiDateLengthsV1Marker

Marker type for DateLengthsV1: “datetime/dangi/datelengths@1”

DangiDateSymbolsV1Marker

Marker type for DateSymbolsV1: “datetime/dangi/datesymbols@1”

DateLengthsV1

Pattern data for dates.

DateSymbolsV1

Symbol data for the months, weekdays, and eras needed to format a date.

Eras

String data for the name, abbreviation, and narrow form of a date’s era.

EthiopianDateLengthsV1Marker

Marker type for DateLengthsV1: “datetime/ethiopic/datelengths@1”

EthiopianDateSymbolsV1Marker

Marker type for DateSymbolsV1: “datetime/ethiopic/datesymbols@1”

GregorianDateLengthsV1Marker

Marker type for DateLengthsV1: “datetime/gregory/datelengths@1”

GregorianDateSymbolsV1Marker

Marker type for DateSymbolsV1: “datetime/gregory/datesymbols@1”

HebrewDateLengthsV1Marker

Marker type for DateLengthsV1: “datetime/hebrew/datelengths@1”

HebrewDateSymbolsV1Marker

Marker type for DateSymbolsV1: “datetime/hebrew/datesymbols@1”

IndianDateLengthsV1Marker

Marker type for DateLengthsV1: “datetime/indian/datelengths@1”

IndianDateSymbolsV1Marker

Marker type for DateSymbolsV1: “datetime/indian/datesymbols@1”

IslamicDateLengthsV1Marker

Marker type for DateLengthsV1: “datetime/islamic/datelengths@1”

IslamicDateSymbolsV1Marker

Marker type for DateSymbolsV1: “datetime/islamic/datesymbols@1”

JapaneseDateLengthsV1Marker

Marker type for DateLengthsV1: “datetime/japanese/datelengths@1”

JapaneseDateSymbolsV1Marker

Marker type for DateSymbolsV1: “datetime/japanese/datesymbols@1”

JapaneseExtendedDateLengthsV1Marker

Marker type for DateLengthsV1: “datetime/japanext/datelengths@1”

JapaneseExtendedDateSymbolsV1Marker

Marker type for DateSymbolsV1: “datetime/japanext/datesymbols@1”

PersianDateLengthsV1Marker

Marker type for DateLengthsV1: “datetime/persian/datelengths@1”

PersianDateSymbolsV1Marker

Marker type for DateSymbolsV1: “datetime/persian/datesymbols@1”

RocDateLengthsV1Marker

Marker type for DateLengthsV1: “datetime/roc/datelengths@1”

RocDateSymbolsV1Marker

Marker type for DateSymbolsV1: “datetime/roc/datesymbols@1”

TimeLengthsV1

Pattern data for times.

TimeLengthsV1Marker

Marker type for TimeLengthsV1: “datetime/timelengths@1”

TimeSymbolsV1

Symbol data for the day periods needed to format a time.

TimeSymbolsV1Marker

Marker type for TimeSymbolsV1: “datetime/timesymbols@1”