icu::datetime::provider

Module fields

Enums representing the fields in a date pattern, including the field’s type, length and symbol.

Structs

DayPeriodULE

ULE type for DayPeriod

DayULE

ULE type for Day

DecimalSecondULE

ULE type for DecimalSecond

Field

A field within a date pattern string, also referred to as a date field.

FieldSymbolULE

ULE type for FieldSymbol

FieldULE

ULE type for Field

HourULE

ULE type for Hour

MonthULE

ULE type for Month

SecondULE

ULE type for Second

TimeZoneULE

ULE type for TimeZone

WeekdayULE

ULE type for Weekday

YearULE

ULE type for Year

Enums

Day

An enum for the possible symbols of a day field in a date pattern.

DayPeriod

An enum for the possible symbols of a day period field in a date pattern.

DecimalSecond

A second field with fractional digits.

Error

An error relating to the field for a date pattern field as a whole.

FieldLength

An enum representing the length of a field within a date or time formatting pattern string.

FieldNumericOverrides

Various numeric overrides for datetime patterns as found in CLDR

FieldSymbol

A field symbol for a date formatting pattern.

Hour

An enum for the possible symbols of an hour field in a date pattern.

LengthError

An error relating to the length of a field within a date pattern.

Month

An enum for the possible symbols of a month field in a date pattern.

Second

An enum for the possible symbols of a second field in a date pattern.

SymbolError

An error relating to the field symbol for a date pattern field.

TimeZone

An enum for the possible symbols of a time zone field in a date pattern.

Week

An enum for the possible symbols of a week field in a date pattern.

Weekday

An enum for the possible symbols of a weekday field in a date pattern.

Year

An enum for the possible symbols of a year field in a date pattern.