icu::datetime::provider

Module pattern

Structured datetime pattern types for datagen and the data provider.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Modules

reference

Reference Pattern implementation intended to follow Unicode specification, and become publicly available for tooling to use for parsing/inspecting/modifying and serialization.

runtime

Runtime Pattern implementation which is optimized for zero-allocation deserialization and high-performance runtime use in DateTimeFormatter.

Enums

CoarseHourCycle

Used to represent either H11/H12, or H23. Skeletons only store these hour cycles as H12 or H23.

GenericPatternItem

An element of a GenericPattern.

PatternError

A low-level pattern parsing error.

PatternItem

An element of a Pattern.

TimeGranularity

The granularity of time represented in a Pattern. Ordered from least granular to most granular for comparison.