icu::datetime::provider::pattern::runtime

Struct GenericPattern

pub struct GenericPattern<'data> {
    pub items: ZeroVec<'data, GenericPatternItem>,
}

A raw, low-level pattern with literals and placeholders.

This is a datetime-specific type designed to be binary-compatible with Pattern. ICU4X developers looking for this sort of type should use the icu_pattern crate instead.

๐ง This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Fields
items: ZeroVec<'data, GenericPatternItem>

The list of GenericPatternItems.

Implementations

impl GenericPattern<'_>

pub fn to_string(&self) -> String

Converts the given value to a String.

Under the hood, this uses an efficient Writeable implementation. However, in order to avoid allocating a string, it is more efficient to use Writeable directly.

impl<'data> GenericPattern<'data>

pub fn combined( self, date: Pattern<'data>, time: Pattern<'data>, ) -> Result<Pattern<'static>, PatternError>

The function allows for creation of new DTF pattern from a generic pattern and replacement patterns.

Examples

use icu::datetime::provider::pattern::runtime::{GenericPattern, Pattern};

let date: Pattern = "y-M-d".parse().expect("Failed to parse pattern");
let time: Pattern = "HH:mm".parse().expect("Failed to parse pattern");

let glue: GenericPattern = "{1} 'at' {0}"
    .parse()
    .expect("Failed to parse generic pattern");
assert_eq!(
    glue.combined(date, time)
        .expect("Failed to combine patterns")
        .to_string(),
    "y-M-d 'at' HH:mm"
);