Struct icu::datetime::provider::Baked

pub struct Baked;

Baked data

๐ง This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. In particular, the `DataProvider` implementations are only guaranteed to match with this version's `*_unstable` providers. Use with caution.