Struct icu::datetime::provider::time_zones::MetazoneSpecificNamesShortV1

pub struct MetazoneSpecificNamesShortV1<'data> {
    pub defaults: ZeroMap2d<'data, MetazoneId, ZoneVariant, str>,
    pub overrides: ZeroMap2d<'data, TimeZoneBcp47Id, ZoneVariant, str>,
}

An ICU4X mapping to the short-form specific metazone names. Specific names include time variants such as โdaylight.โ See CLDR-JSON timeZoneNames.json for more context.

๐ง This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Fields
defaults: ZeroMap2d<'data, MetazoneId, ZoneVariant, str>

The default mapping between metazone id and localized metazone name.

overrides: ZeroMap2d<'data, TimeZoneBcp47Id, ZoneVariant, str>

The override mapping between timezone id and localized metazone name.