icu::datetime::scaffold

Trait AllAnyCalendarFormattingDataMarkers

pub trait AllAnyCalendarFormattingDataMarkers<FSet>:
    DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Year as CalMarkers<YearNamesV1>>::Buddhist>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Year as CalMarkers<YearNamesV1>>::Chinese>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Year as CalMarkers<YearNamesV1>>::Coptic>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Year as CalMarkers<YearNamesV1>>::Dangi>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Year as CalMarkers<YearNamesV1>>::Ethiopian>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Year as CalMarkers<YearNamesV1>>::Gregorian>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Year as CalMarkers<YearNamesV1>>::Hebrew>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Year as CalMarkers<YearNamesV1>>::Indian>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Year as CalMarkers<YearNamesV1>>::Hijri>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Year as CalMarkers<YearNamesV1>>::Japanese>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Year as CalMarkers<YearNamesV1>>::Persian>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Year as CalMarkers<YearNamesV1>>::Roc>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Month as CalMarkers<MonthNamesV1>>::Buddhist>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Month as CalMarkers<MonthNamesV1>>::Chinese>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Month as CalMarkers<MonthNamesV1>>::Coptic>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Month as CalMarkers<MonthNamesV1>>::Dangi>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Month as CalMarkers<MonthNamesV1>>::Ethiopian>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Month as CalMarkers<MonthNamesV1>>::Gregorian>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Month as CalMarkers<MonthNamesV1>>::Hebrew>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Month as CalMarkers<MonthNamesV1>>::Indian>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Month as CalMarkers<MonthNamesV1>>::Hijri>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Month as CalMarkers<MonthNamesV1>>::Japanese>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Month as CalMarkers<MonthNamesV1>>::Persian>
    + DataProvider<<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::Month as CalMarkers<MonthNamesV1>>::Roc>
    + DataProvider<<<FSet as DateTimeMarkers>::D as DateDataMarkers>::WeekdayNamesV1>
    + DataProvider<<<FSet as DateTimeMarkers>::T as TimeMarkers>::DayPeriodNamesV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::EssentialsV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::LocationsV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::LocationsRootV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::ExemplarCitiesV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::ExemplarCitiesRootV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::GenericLongV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::GenericShortV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::StandardLongV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::SpecificLongV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::SpecificShortV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::MetazonePeriodV1>
    + AllAnyCalendarPatternDataMarkers<FSet>
where
    FSet: DateTimeMarkers,
    <FSet as DateTimeMarkers>::D: DateDataMarkers,
    <FSet as DateTimeMarkers>::T: TimeMarkers,
    <FSet as DateTimeMarkers>::Z: ZoneMarkers,
{ }

Trait to consolidate data provider markers defined by this crate for datetime formatting with any calendar.

This trait is implemented on all providers that support datetime formatting, including crate::provider::Baked.