icu::datetime::scaffold

Trait AllFixedCalendarFormattingDataMarkers

pub trait AllFixedCalendarFormattingDataMarkers<C, FSet>:
    DataProvider<<<FSet as DateTimeMarkers>::D as TypedDateDataMarkers<C>>::YearNamesV1>
    + DataProvider<<<FSet as DateTimeMarkers>::D as TypedDateDataMarkers<C>>::MonthNamesV1>
    + DataProvider<<<FSet as DateTimeMarkers>::D as TypedDateDataMarkers<C>>::WeekdayNamesV1>
    + DataProvider<<<FSet as DateTimeMarkers>::T as TimeMarkers>::DayPeriodNamesV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::EssentialsV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::LocationsV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::LocationsRootV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::ExemplarCitiesV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::ExemplarCitiesRootV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::GenericLongV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::GenericShortV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::StandardLongV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::SpecificLongV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::SpecificShortV1>
    + DataProvider<<<FSet as DateTimeMarkers>::Z as ZoneMarkers>::MetazonePeriodV1>
    + AllFixedCalendarPatternDataMarkers<C, FSet>
where
    C: CldrCalendar,
    FSet: DateTimeMarkers,
    <FSet as DateTimeMarkers>::D: TypedDateDataMarkers<C>,
    <FSet as DateTimeMarkers>::T: TimeMarkers,
    <FSet as DateTimeMarkers>::Z: ZoneMarkers,
{ }

Trait to consolidate data provider markers defined by this crate for datetime formatting with a fixed calendar.

This trait is implemented on all providers that support datetime formatting, including crate::provider::Baked.