icu::datetime::scaffold

Trait AllInputMarkers

pub trait AllInputMarkers<R>:
    GetField<<<R as DateTimeMarkers>::D as DateInputMarkers>::YearInput>
    + GetField<<<R as DateTimeMarkers>::D as DateInputMarkers>::MonthInput>
    + GetField<<<R as DateTimeMarkers>::D as DateInputMarkers>::DayOfMonthInput>
    + GetField<<<R as DateTimeMarkers>::D as DateInputMarkers>::DayOfWeekInput>
    + GetField<<<R as DateTimeMarkers>::D as DateInputMarkers>::DayOfYearInput>
    + GetField<<<R as DateTimeMarkers>::T as TimeMarkers>::HourInput>
    + GetField<<<R as DateTimeMarkers>::T as TimeMarkers>::MinuteInput>
    + GetField<<<R as DateTimeMarkers>::T as TimeMarkers>::SecondInput>
    + GetField<<<R as DateTimeMarkers>::T as TimeMarkers>::NanosecondInput>
    + GetField<<<R as DateTimeMarkers>::Z as ZoneMarkers>::TimeZoneIdInput>
    + GetField<<<R as DateTimeMarkers>::Z as ZoneMarkers>::TimeZoneOffsetInput>
    + GetField<<<R as DateTimeMarkers>::Z as ZoneMarkers>::TimeZoneVariantInput>
    + GetField<<<R as DateTimeMarkers>::Z as ZoneMarkers>::TimeZoneNameTimestampInput>
where
    R: DateTimeMarkers,
    <R as DateTimeMarkers>::D: DateInputMarkers,
    <R as DateTimeMarkers>::T: TimeMarkers,
    <R as DateTimeMarkers>::Z: ZoneMarkers,
{ }

Trait implemented on formattable datetime types.

This trait allows for only those types compatible with a particular field set to be passed as arguments to the formatting function for that field set. For example, this trait prevents Time from being passed to a formatter parameterized with fieldsets::YMD.

The following types implement this trait:

Date

Time

DateTime

ZonedDateTime

UtcOffset

TimeZoneInfo