icu::datetime::scaffold

Trait CalMarkers

pub trait CalMarkers<M>: UnstableSealed
where
    M: DynamicDataMarker,
{
    type Buddhist: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>;
    type Chinese: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>;
    type Coptic: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>;
    type Dangi: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>;
    type Ethiopian: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>;
    type Gregorian: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>;
    type Hebrew: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>;
    type Indian: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>;
    type Hijri: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>;
    type Japanese: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>;
    type Persian: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>;
    type Roc: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>;
}

A collection of marker types associated with all formattable calendars.

This is used to group together the calendar-specific marker types that produce a common DynamicDataMarker. For example, this trait can be implemented for YearNamesV1.

This trait serves as a building block for a cross-calendar BoundDataProvider.

๐ง This trait is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. Do not implement this trait in userland unless you are prepared for things to occasionally break.

Required Associated Types

type Buddhist: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>

The type for a Buddhist calendar

type Chinese: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>

The type for a Chinese calendar

type Coptic: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>

The type for a Coptic calendar

type Dangi: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>

The type for a Dangi calendar

type Ethiopian: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>

The type for an Ethiopian calendar (either era style)

type Gregorian: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>

The type for a Gregorian calendar

type Hebrew: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>

The type for a Hebrew calendar

type Indian: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>

The type for a Indian calendar

type Hijri: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>

The type for Hirji calendars

type Japanese: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>

The type for a Japanese calendar

type Persian: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>

The type for a Persian calendar

type Roc: DataMarker<DataStruct = <M as DynamicDataMarker>::DataStruct>

The type for a Roc calendar