icu::datetime::scaffold

Trait CldrCalendar

pub trait CldrCalendar: UnstableSealed {
    type YearNamesV1: DataMarker<DataStruct = YearNames<'static>>;
    type MonthNamesV1: DataMarker<DataStruct = MonthNames<'static>>;
    type SkeletaV1: DataMarker<DataStruct = PackedPatterns<'static>>;
}

A calendar that can be found in CLDR.

New implementors of this trait will likely also wish to modify get_era_code_map() in the CLDR transformer to support any new era maps.

๐ซ This trait is sealed; it cannot be implemented by user code. If an API requests an item that implements this trait, please consider using a type from the implementors listed below.

Required Associated Types

type YearNamesV1: DataMarker<DataStruct = YearNames<'static>>

The data marker for loading year symbols for this calendar.

type MonthNamesV1: DataMarker<DataStruct = MonthNames<'static>>

The data marker for loading month symbols for this calendar.

type SkeletaV1: DataMarker<DataStruct = PackedPatterns<'static>>

The data marker for loading skeleton patterns for this calendar.