icu::datetime::scaffold

Trait DateTimeMarkers

pub trait DateTimeMarkers: UnstableSealed + DateTimeNamesMarker {
    type D;
    type T;
    type Z;
    type GluePatternV1: DataMarker<DataStruct = GluePattern<'static>>;
}

A trait associating constants and types implementing various other traits required for datetime formatting.

This is a sealed trait implemented on field set markers.

๐ง This trait is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. Do not implement this trait in userland unless you are prepared for things to occasionally break.

Required Associated Types

type D

Associated types for date formatting.

Should implement DateDataMarkers, TypedDateDataMarkers, and DateInputMarkers.

type T

Associated types for time formatting.

Should implement TimeMarkers.

type Z

Associated types for time zone formatting.

Should implement ZoneMarkers.

type GluePatternV1: DataMarker<DataStruct = GluePattern<'static>>

Marker for loading the date/time glue pattern.