icu::datetime::scaffold

Trait DateTimeNamesMarker

pub trait DateTimeNamesMarker: UnstableSealed {

    type YearNames: NamesContainer<YearNamesV1, YearNameLength>;
    type MonthNames: NamesContainer<MonthNamesV1, MonthNameLength>;
    type WeekdayNames: NamesContainer<DatetimeNamesWeekdayV1, WeekdayNameLength>;
    type DayPeriodNames: NamesContainer<DatetimeNamesDayperiodV1, DayPeriodNameLength>;
    type ZoneEssentials: NamesContainer<TimezoneNamesEssentialsV1, ()>;
    type ZoneLocations: NamesContainer<TimezoneNamesLocationsOverrideV1, ()>;
    type ZoneLocationsRoot: NamesContainer<TimezoneNamesLocationsRootV1, ()>;
    type ZoneExemplars: NamesContainer<TimezoneNamesCitiesOverrideV1, ()>;
    type ZoneExemplarsRoot: NamesContainer<TimezoneNamesCitiesRootV1, ()>;
    type ZoneGenericLong: NamesContainer<TimezoneNamesGenericLongV1, ()>;
    type ZoneGenericShort: NamesContainer<TimezoneNamesGenericShortV1, ()>;
    type ZoneStandardLong: NamesContainer<TimezoneNamesStandardLongV1, ()>;
    type ZoneSpecificLong: NamesContainer<TimezoneNamesSpecificLongV1, ()>;
    type ZoneSpecificShort: NamesContainer<TimezoneNamesSpecificShortV1, ()>;
    type MetazoneLookup: NamesContainer<TimezoneMetazonePeriodsV1, ()>;

}

Trait for a type that owns datetime names data, usually in the form of data payloads.

This trait allows for types that contain data for some but not all types of datetime names, allowing for reduced stack size. For example, a type could contain year and month names but not weekday, day period, or time zone names.

๐ง This trait is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. Do not implement this trait in userland unless you are prepared for things to occasionally break.

Required Associated Types

type YearNames: NamesContainer<YearNamesV1, YearNameLength>

type MonthNames: NamesContainer<MonthNamesV1, MonthNameLength>

type WeekdayNames: NamesContainer<DatetimeNamesWeekdayV1, WeekdayNameLength>

type DayPeriodNames: NamesContainer<DatetimeNamesDayperiodV1, DayPeriodNameLength>

type ZoneEssentials: NamesContainer<TimezoneNamesEssentialsV1, ()>

type ZoneLocations: NamesContainer<TimezoneNamesLocationsOverrideV1, ()>

type ZoneLocationsRoot: NamesContainer<TimezoneNamesLocationsRootV1, ()>

type ZoneExemplars: NamesContainer<TimezoneNamesCitiesOverrideV1, ()>

type ZoneExemplarsRoot: NamesContainer<TimezoneNamesCitiesRootV1, ()>

type ZoneGenericLong: NamesContainer<TimezoneNamesGenericLongV1, ()>

type ZoneGenericShort: NamesContainer<TimezoneNamesGenericShortV1, ()>

type ZoneStandardLong: NamesContainer<TimezoneNamesStandardLongV1, ()>

type ZoneSpecificLong: NamesContainer<TimezoneNamesSpecificLongV1, ()>

type ZoneSpecificShort: NamesContainer<TimezoneNamesSpecificShortV1, ()>

type MetazoneLookup: NamesContainer<TimezoneMetazonePeriodsV1, ()>