icu::datetime::scaffold

Trait GetField

pub trait GetField<T>: UnstableSealed {
    // Required method
    fn get_field(&self) -> T;
}

A type that can return a certain field T.

This is used as a bound on various datetime functions.

๐ง This trait is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. Do not implement this trait in userland unless you are prepared for things to occasionally break.

Required Methods

fn get_field(&self) -> T

Returns the value of this traitโs field T.