icu::datetime::scaffold

Trait InSameCalendar

pub trait InSameCalendar {
    // Required method
    fn check_any_calendar_kind(
        &self,
        any_calendar_kind: AnyCalendarKind,
    ) -> Result<(), MismatchedCalendarError>;
}

An input that may be associated with a specific runtime calendar.

Required Methods

fn check_any_calendar_kind( &self, any_calendar_kind: AnyCalendarKind, ) -> Result<(), MismatchedCalendarError>

Checks whether this type is compatible with the given calendar.

Types that are agnostic to calendar systems should return Ok(()).