icu::datetime::scaffold

Trait TypedDateDataMarkers

pub trait TypedDateDataMarkers<C>: UnstableSealed {
    type DateSkeletonPatternsV1: DataMarker<DataStruct = PackedPatterns<'static>>;
    type YearNamesV1: DataMarker<DataStruct = YearNames<'static>>;
    type MonthNamesV1: DataMarker<DataStruct = MonthNames<'static>>;
    type WeekdayNamesV1: DataMarker<DataStruct = LinearNames<'static>>;
}

A trait associating types for date formatting in a specific calendar (data markers only).

This is a sealed trait implemented on field set markers.

๐ง This trait is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. Do not implement this trait in userland unless you are prepared for things to occasionally break.

Required Associated Types

type DateSkeletonPatternsV1: DataMarker<DataStruct = PackedPatterns<'static>>

Marker for loading date skeleton patterns.

type YearNamesV1: DataMarker<DataStruct = YearNames<'static>>

Marker for loading year names.

type MonthNamesV1: DataMarker<DataStruct = MonthNames<'static>>

Marker for loading month names.

type WeekdayNamesV1: DataMarker<DataStruct = LinearNames<'static>>

Marker for loading weekday names.