icu::datetime::scaffold

Trait ZoneMarkers

pub trait ZoneMarkers: UnstableSealed {

    type TimeZoneIdInput: IntoOption<TimeZone>;
    type TimeZoneOffsetInput: IntoOption<UtcOffset>;
    type TimeZoneVariantInput: IntoOption<TimeZoneVariant>;
    type TimeZoneNameTimestampInput: IntoOption<ZoneNameTimestamp>;
    type EssentialsV1: DataMarker<DataStruct = TimeZoneEssentials<'static>>;
    type LocationsV1: DataMarker<DataStruct = Locations<'static>>;
    type LocationsRootV1: DataMarker<DataStruct = Locations<'static>>;
    type ExemplarCitiesV1: DataMarker<DataStruct = ExemplarCities<'static>>;
    type ExemplarCitiesRootV1: DataMarker<DataStruct = ExemplarCities<'static>>;
    type GenericLongV1: DataMarker<DataStruct = MetazoneGenericNames<'static>>;
    type GenericShortV1: DataMarker<DataStruct = MetazoneGenericNames<'static>>;
    type StandardLongV1: DataMarker<DataStruct = MetazoneGenericNames<'static>>;
    type SpecificLongV1: DataMarker<DataStruct = MetazoneSpecificNames<'static>>;
    type SpecificShortV1: DataMarker<DataStruct = MetazoneSpecificNames<'static>>;
    type MetazonePeriodV1: DataMarker<DataStruct = MetazonePeriod<'static>>;

}

A trait associating types for time zone formatting (input types and data markers).

This is a sealed trait implemented on field set markers.

๐ง This trait is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. Do not implement this trait in userland unless you are prepared for things to occasionally break.

Required Associated Types

type TimeZoneIdInput: IntoOption<TimeZone>

Marker for resolving the time zone id input field.

type TimeZoneOffsetInput: IntoOption<UtcOffset>

Marker for resolving the time zone offset input field.

type TimeZoneVariantInput: IntoOption<TimeZoneVariant>

Marker for resolving the time zone variant input field.

type TimeZoneNameTimestampInput: IntoOption<ZoneNameTimestamp>

Marker for resolving the time zone non-location display names, which depend on the datetime.

type EssentialsV1: DataMarker<DataStruct = TimeZoneEssentials<'static>>

Marker for loading core time zone data.

type LocationsV1: DataMarker<DataStruct = Locations<'static>>

Marker for loading location names for time zone formatting

type LocationsRootV1: DataMarker<DataStruct = Locations<'static>>

Marker for loading location names for time zone formatting

type ExemplarCitiesV1: DataMarker<DataStruct = ExemplarCities<'static>>

Marker for loading exemplar city names for time zone formatting

type ExemplarCitiesRootV1: DataMarker<DataStruct = ExemplarCities<'static>>

Marker for loading exemplar city names for time zone formatting

type GenericLongV1: DataMarker<DataStruct = MetazoneGenericNames<'static>>

Marker for loading generic long time zone names.

type GenericShortV1: DataMarker<DataStruct = MetazoneGenericNames<'static>>

Marker for loading generic short time zone names.

type StandardLongV1: DataMarker<DataStruct = MetazoneGenericNames<'static>>

Marker for loading standard long time zone names.

type SpecificLongV1: DataMarker<DataStruct = MetazoneSpecificNames<'static>>

Marker for loading specific long time zone names.

type SpecificShortV1: DataMarker<DataStruct = MetazoneSpecificNames<'static>>

Marker for loading generic short time zone names.

type MetazonePeriodV1: DataMarker<DataStruct = MetazonePeriod<'static>>

Marker for loading metazone periods.