Module icu::datetime::skeleton

Skeletons are used for pattern matching. See the Skeleton struct for more information.

Structs

AvailableFormatPattern
The AvailableFormatPattern represents a specific pattern that is available for a given locale. A Skeleton is used to match against to find the best pattern.

Skeleton
A Skeleton is used to represent what types of Fields are present in a Pattern. The ordering of the Skeleton's Fields have no bearing on the ordering of the Fields and Literals in the Pattern.

Enums

BestSkeleton
According to the UTS 35 skeleton matching algorithm there will be a guaranteed match for a skeleton. However, with this initial implementation, there is no attempt to add on missing fields. This enum encodes the variants for the current search for a best skeleton.

SkeletonError

Functions

get_available_format_patterns

get_best_available_format_pattern
A partial implementation of the UTS 35 skeleton matching algorithm.