icu::datetime

Struct FormattedDateTime

pub struct FormattedDateTime<'a> { /* private fields */ }

An intermediate type during a datetime formatting operation.

Not intended to be stored: convert to a string first.

Implementations

impl FormattedDateTime<'_>

pub fn to_string(&self) -> String

Converts the given value to a String.

Under the hood, this uses an efficient Writeable implementation. However, in order to avoid allocating a string, it is more efficient to use Writeable directly.

impl FormattedDateTime<'_>

pub fn pattern(&self) -> DateTimePattern

Gets the pattern used in this formatted value.

From the pattern, one can check the properties of the included components, such as the hour cycle being used for formatting. See DateTimePattern.