Struct icu::datetime::TimeFormatter

pub struct TimeFormatter(_);

TimeFormatter is a structure of the icu_datetime component that provides time formatting only. When constructed, it uses data from the data provider, selected locale and provided preferences to collect all data necessary to format any time into that locale.

For that reason, one should think of the process of formatting a time in two steps - first, a computational heavy construction of TimeFormatter, and then fast formatting of DateTimeInput data using the instance.

Examples

use icu::calendar::DateTime;
use icu::datetime::{options::length::Time, TimeFormatter};
use icu::locid::locale;
use writeable::assert_writeable_eq;

let tf = TimeFormatter::try_new_with_length_unstable(
    &icu_testdata::unstable(),
    &locale!("en").into(),
    Time::Short,
)
.expect("Failed to create TimeFormatter instance.");

let datetime = DateTime::try_new_gregorian_datetime(2020, 9, 1, 12, 34, 28)
    .expect("Failed to construct DateTime.");

assert_writeable_eq!(tf.format(&datetime), "12:34โฏPM");

This model replicates that of ICU and ECMA402.

Implementations

impl TimeFormatter

pub fn try_new_with_length_unstable<D>( data_provider: &D, locale: &DataLocale, length: Time ) -> Result<TimeFormatter, DateTimeError>where D: DataProvider<TimeLengthsV1Marker> + DataProvider<TimeSymbolsV1Marker> + DataProvider<DecimalSymbolsV1Marker> + ?Sized,

Constructor that takes a selected locale, reference to a data provider and a list of preferences, then collects all data necessary to format date and time values into the given locale, using the short style.

๐ Help choosing a constructor

โ๏ธ The bounds on this function may change over time, including in SemVer minor releases.

Examples

use icu::datetime::{options::length::Time, TimeFormatter};
use icu::locid::locale;

TimeFormatter::try_new_with_length_unstable(
    &icu_testdata::unstable(),
    &locale!("en").into(),
    Time::Short,
)
.unwrap();

pub fn try_new_with_length_with_any_provider( provider: &impl AnyProvider, locale: &DataLocale, length: Time ) -> Result<TimeFormatter, DateTimeError>

Creates a new instance using an AnyProvider.

For details on the behavior of this function, see: Self::try_new_with_length_unstable

๐ Help choosing a constructor

pub fn format<T, 'l>(&'l self, value: &T) -> FormattedDateTime<'l>where T: IsoTimeInput,

Takes a IsoTimeInput implementer and returns an instance of a FormattedDateTime that contains all information necessary to display a formatted date and operate on it.

Examples

use icu::calendar::DateTime;
use icu::datetime::{options::length::Time, TimeFormatter};
use icu::locid::locale;
use writeable::assert_writeable_eq;

let tf = TimeFormatter::try_new_with_length_unstable(
    &icu_testdata::unstable(),
    &locale!("en").into(),
    Time::Short,
)
.expect("Failed to create TimeFormatter instance.");

let datetime = DateTime::try_new_gregorian_datetime(2020, 9, 1, 12, 34, 28)
    .expect("Failed to construct DateTime.");

assert_writeable_eq!(tf.format(&datetime), "12:34โฏPM");

pub fn format_to_string(&self, value: &impl IsoTimeInput) -> String

Takes a IsoTimeInput implementer and returns it formatted as a string.

Examples

use icu::calendar::DateTime;
use icu::datetime::{options::length::Time, TimeFormatter};
use icu::locid::locale;
let tf = TimeFormatter::try_new_with_length_unstable(
    &icu_testdata::unstable(),
    &locale!("en").into(),
    Time::Short,
)
.expect("Failed to create TimeFormatter instance.");

let datetime = DateTime::try_new_gregorian_datetime(2020, 9, 1, 12, 34, 28)
    .expect("Failed to construct DateTime.");

assert_eq!(tf.format_to_string(&datetime), "12:34โฏPM");