Struct icu::datetime::ZonedDateTimeFormat

pub struct ZonedDateTimeFormat<C>(_, _);

The composition of DateTimeFormat and TimeZoneFormat.

ZonedDateTimeFormat uses data from the [data provider]s, the selected Locale, and the provided pattern to collect all data necessary to format a datetime with time zones into that locale.

The various pattern symbols specified in UTS-35 require different sets of data for formatting. As such, TimeZoneFormat will pull in only the resources it needs to format that pattern that is derived from the provided DateTimeFormatOptions.

For that reason, one should think of the process of formatting a zoned datetime in two steps: first, a computationally heavy construction of ZonedDateTimeFormat, and then fast formatting of the data using the instance.

Examples

use icu::locid::Locale;
use icu::locid::langid;
use icu::calendar::Gregorian;
use icu::datetime::{ZonedDateTimeFormat, options::length};
use icu::datetime::mock::zoned_datetime::MockZonedDateTime;
use icu_provider::inv::InvariantDataProvider;

let locale: Locale = langid!("en").into();

let date_provider = InvariantDataProvider;
let zone_provider = InvariantDataProvider;
let plural_provider = InvariantDataProvider;

let options = length::Bag {
    date: Some(length::Date::Medium),
    time: Some(length::Time::Short),
    ..Default::default()
};
let zdtf = ZonedDateTimeFormat::<Gregorian>::try_new(locale, &date_provider, &zone_provider, &plural_provider, &options.into())
    .expect("Failed to create DateTimeFormat instance.");

let zoned_datetime: MockZonedDateTime = "2021-04-08T16:12:37.000-07:00"
    .parse()
    .expect("Failed to parse zoned datetime");

let value = zdtf.format_to_string(&zoned_datetime);

Implementations

impl<C> ZonedDateTimeFormat<C> where
C: CldrCalendar,

pub fn try_new<L, DP, ZP, PP>(
locale: L,
date_provider: &DP,
zone_provider: &ZP,
plural_provider: &PP,
options: &DateTimeFormatOptions
) -> Result<ZonedDateTimeFormat<C>, DateTimeFormatError> where
L: Into<Locale>,
DP: ResourceProvider<DateSymbolsV1Marker> + ResourceProvider<DatePatternsV1Marker> + ResourceProvider<DateSkeletonPatternsV1Marker> + ResourceProvider<WeekDataV1Marker>,
ZP: ResourceProvider<TimeZoneFormatsV1Marker> + ResourceProvider<ExemplarCitiesV1Marker> + ResourceProvider<MetaZoneGenericNamesLongV1Marker> + ResourceProvider<MetaZoneGenericNamesShortV1Marker> + ResourceProvider<MetaZoneSpecificNamesLongV1Marker> + ResourceProvider<MetaZoneSpecificNamesShortV1Marker> + ?Sized,
PP: ResourceProvider<OrdinalV1Marker>,

Constructor that takes a selected Locale, a reference to a data provider for dates, a data provider for time zones, and a list of DateTimeFormatOptions. It collects all data necessary to format zoned datetime values into the given locale.

Examples

use icu::locid::Locale;
use icu::locid::langid;
use icu::calendar::Gregorian;
use icu::datetime::{ZonedDateTimeFormat, DateTimeFormatOptions};
use icu::datetime::mock::zoned_datetime::MockZonedDateTime;
use icu_provider::inv::InvariantDataProvider;

let locale: Locale = langid!("en").into();

let date_provider = InvariantDataProvider;
let zone_provider = InvariantDataProvider;
let plural_provider = InvariantDataProvider;

let options = DateTimeFormatOptions::default();

let zdtf = ZonedDateTimeFormat::<Gregorian>::try_new(locale, &date_provider, &zone_provider, &plural_provider, &options);

assert_eq!(zdtf.is_ok(), true);

pub fn format<T>(&'l self, value: &'l T) -> FormattedZonedDateTime<'l, T> where
T: ZonedDateTimeInput,

Takes a ZonedDateTimeInput implementer and returns an instance of a FormattedZonedDateTime that contains all information necessary to display a formatted zoned datetime and operate on it.

Examples

use icu::locid::Locale;
use icu::locid::langid;
use icu::calendar::Gregorian;
use icu::datetime::{ZonedDateTimeFormat, DateTimeFormatOptions};
use icu::datetime::mock::zoned_datetime::MockZonedDateTime;
use icu_provider::inv::InvariantDataProvider;
let zdtf = ZonedDateTimeFormat::<Gregorian>::try_new(locale, &date_provider, &zone_provider, &plural_provider, &options)
    .expect("Failed to create ZonedDateTimeFormat instance.");

let zoned_datetime: MockZonedDateTime = "2021-04-08T16:12:37.000-07:00"
    .parse()
    .expect("Failed to parse zoned datetime");

let formatted_date = zdtf.format(&zoned_datetime);

let _ = format!("Date: {}", formatted_date);

At the moment, there’s little value in using that over one of the other format methods, but FormattedZonedDateTime will grow with methods for iterating over fields, extracting information about formatted date and so on.

pub fn format_to_write(
&self,
w: &mut impl Write,
value: &impl ZonedDateTimeInput
) -> Result<(), Error>

Takes a mutable reference to anything that implements the Write trait and a ZonedDateTimeInput implementer, then populates the buffer with a formatted value.

Examples

use icu::locid::Locale;
use icu::locid::langid;
use icu::calendar::Gregorian;
use icu::datetime::{ZonedDateTimeFormat, DateTimeFormatOptions};
use icu::datetime::mock::zoned_datetime::MockZonedDateTime;
use icu_provider::inv::InvariantDataProvider;
let zdtf = ZonedDateTimeFormat::<Gregorian>::try_new(locale, &date_provider, &zone_provider, &plural_provider, &options.into())
    .expect("Failed to create ZonedDateTimeFormat instance.");

let zoned_datetime: MockZonedDateTime = "2021-04-08T16:12:37.000-07:00"
    .parse()
    .expect("Failed to parse zoned datetime");

let mut buffer = String::new();
zdtf.format_to_write(&mut buffer, &zoned_datetime)
    .expect("Failed to write to a buffer.");

let _ = format!("Date: {}", buffer);

pub fn format_to_string(&self, value: &impl ZonedDateTimeInput) -> String

Takes a ZonedDateTimeInput implementer and returns it formatted as a string.

Examples

use icu::locid::Locale;
use icu::locid::langid;
use icu::calendar::Gregorian;
use icu::datetime::{ZonedDateTimeFormat, DateTimeFormatOptions};
use icu::datetime::mock::zoned_datetime::MockZonedDateTime;
use icu_provider::inv::InvariantDataProvider;
let zdtf = ZonedDateTimeFormat::<Gregorian>::try_new(locale, &date_provider, &zone_provider, &plural_provider, &options.into())
    .expect("Failed to create ZonedDateTimeFormat instance.");

let zoned_datetime: MockZonedDateTime = "2021-04-08T16:12:37.000-07:00"
    .parse()
    .expect("Failed to parse zoned datetime");

let _ = zdtf.format_to_string(&zoned_datetime);