Trait icu::datetime::CldrCalendar

pub trait CldrCalendar {
    type DateSymbolsV1Marker: KeyedDataMarker
    where
        <Self::DateSymbolsV1Marker as DataMarker>::Yokeable == DateSymbolsV1<'static>;
    type DateLengthsV1Marker: KeyedDataMarker
    where
        <Self::DateLengthsV1Marker as DataMarker>::Yokeable == DateLengthsV1<'static>;

    const DEFAULT_BCP_47_IDENTIFIER: Value;
    fn is_identifier_allowed_for_calendar(value: &Value) -> bool { ... }
}

A calendar that can be found in CLDR

New implementors of this trait will likely also wish to modify get_era_code_map() in the CLDR transformer to support any new era maps.

Associated Types

type DateSymbolsV1Marker: KeyedDataMarker
where
<Self::DateSymbolsV1Marker as DataMarker>::Yokeable == DateSymbolsV1<'static>

The data marker for loading symbols for this calendar.

type DateLengthsV1Marker: KeyedDataMarker
where
<Self::DateLengthsV1Marker as DataMarker>::Yokeable == DateLengthsV1<'static>

The data marker for loading length-patterns for this calendar.

Associated Constants

const DEFAULT_BCP_47_IDENTIFIER: Value

The Unicode BCP 47 identifier for the calendar If multiple BCP 47 identifiers work, this should be the default one when no others are provided

Provided methods

fn is_identifier_allowed_for_calendar(value: &Value) -> bool

Checks if a given BCP 47 identifier is allowed to be used with this calendar

By default, just checks against DEFAULT_BCP_47_IDENTIFIER