icu::datetime::unchecked

Struct DateTimeInputUnchecked

#[non_exhaustive]
pub struct DateTimeInputUnchecked { /* private fields */ }

An input bag with all possible datetime input fields.

Each input field may or may not be required, depending on the field set and the options.

Implementations

impl DateTimeInputUnchecked

pub fn set_date_fields_unchecked<C, A>(&mut self, date_in_calendar: Date<A>)
where C: Calendar, A: AsCalendar<Calendar = C>,

Sets all fields from a Date input.

This method does not check the calendar of the date! The caller is responsible for making sure the calendar matches the formatter.

pub fn set_time_fields(&mut self, time: Time)

Sets all fields from a Time input.

pub fn set_time_zone_utc_offset(&mut self, utc_offset: UtcOffset)

Sets the time zone UTC offset.

pub fn set_time_zone_id(&mut self, id: TimeZone)

Sets the time zone ID.

pub fn set_time_zone_name_timestamp(&mut self, timestamp: ZoneNameTimestamp)

Sets the local time for time zone name resolution.

pub fn set_time_zone_variant(&mut self, _zone_variant: TimeZoneVariant)

๐Deprecated

No-op