icu

Crate decimal

Formatting basic decimal numbers.

This module is published as its own crate (icu_decimal) and as part of the icu crate. See the latter for more details on the ICU4X project.

Support for currencies, measurement units, and compact notation is planned. To track progress, follow icu4x#275.

Examples

Format a number with Bangla digits

use icu::decimal::input::Decimal;
use icu::decimal::DecimalFormatter;
use icu::locale::locale;
use writeable::assert_writeable_eq;

let formatter =
    DecimalFormatter::try_new(locale!("bn").into(), Default::default())
        .expect("locale should be present");

let decimal = Decimal::from(1000007);

assert_writeable_eq!(formatter.format(&decimal), "১০,০০,০০৭");

Format a number with digits after the decimal separator

use icu::decimal::input::Decimal;
use icu::decimal::DecimalFormatter;
use icu::locale::Locale;
use writeable::assert_writeable_eq;

let formatter =
    DecimalFormatter::try_new(Default::default(), Default::default())
        .expect("locale should be present");

let decimal = {
    let mut decimal = Decimal::from(200050);
    decimal.multiply_pow10(-2);
    decimal
};

assert_writeable_eq!(formatter.format(&decimal), "2,000.50");

Format a number using an alternative numbering system

Numbering systems specified in the -u-nu subtag will be followed.

use icu::decimal::input::Decimal;
use icu::decimal::DecimalFormatter;
use icu::locale::locale;
use writeable::assert_writeable_eq;

let formatter = DecimalFormatter::try_new(
    locale!("th-u-nu-thai").into(),
    Default::default(),
)
.expect("locale should be present");

let decimal = Decimal::from(1000007);

assert_writeable_eq!(formatter.format(&decimal), "๑,๐๐๐,๐๐๗");

Modules

input

Types that can be fed to DecimalFormatter and their utilities

options

Options for DecimalFormatter.

parts

Parts of a formatted decimal.

preferences

Locale preferences used by this crate

provider

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

Structs

DecimalFormatter

A formatter for Decimal, rendering decimal digits in an i18n-friendly way.

DecimalFormatterPreferences

The preferences for fixed decimal formatting.

FormattedDecimal

An intermediate structure returned by DecimalFormatter. Use Writeable to render the formatted decimal to a string or buffer.