Crate icu::displaynames

🚧 [Experimental] Display names for regions

This module is published as its own crate (icu_displaynames) and as part of the icu crate. See the latter for more details on the ICU4X project.

🚧 This code is experimental; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. It can be enabled with the "experimental" feature of the icu meta-crate. Use with caution.

Modules

displaynames

This module contains types and implementations for the Displaynames component.

options

Options for DisplayNames.

provider

Data provider struct definitions for this ICU4X component.

Structs

DisplayNames

Lookup of the terrritory display names by region code.

DisplayNamesOptions

A bag of options defining how region codes will be translated by DisplayNames.

Enums

Fallback

An enum for fallback return when the system does not have the requested display name.

LanguageDisplay

An enum for the language display kind.

Style

An enum for formatting style.