icu

Crate list

Formatting lists in a locale-sensitive way.

This module is published as its own crate (icu_list) and as part of the icu crate. See the latter for more details on the ICU4X project.

Examples

Formatting and lists in Spanish

let list_formatter = ListFormatter::try_new_and(
    locale!("es").into(),
    ListFormatterOptions::default().with_length(ListLength::Wide),
)
.expect("locale should be present");

assert_writeable_eq!(
    list_formatter.format(["Espa침a", "Suiza"].iter()),
    "Espa침a y Suiza",
);

// The Spanish 'y' sometimes becomes an 'e':
assert_writeable_eq!(
    list_formatter.format(["Espa침a", "Suiza", "Italia"].iter()),
    "Espa침a, Suiza e Italia",
);

Formatting or lists in Thai

let list_formatter = ListFormatter::try_new_or(
    locale!("th").into(),
    ListFormatterOptions::default().with_length(ListLength::Short),
)
.expect("locale should be present");

// We can use any Writeables as inputs
assert_writeable_eq!(list_formatter.format(1..=3), "1, 2 舡舡舡舡 3",);

Formatting unit lists in English

let list_formatter = ListFormatter::try_new_unit(
    locale!("en").into(),
    ListFormatterOptions::default().with_length(ListLength::Wide),
)
.expect("locale should be present");

assert_writeable_eq!(
    list_formatter.format(["1ft", "2in"].iter()),
    "1ft, 2in",
);

Note: this last example is not fully internationalized. See icu4x/2192 for full unit handling.

Modules

options

Options used by types in this crate

parts

The Parts used by ListFormatter.

provider

游뚾 [Unstable] Data provider struct definitions for this ICU4X component.

Structs

FormattedList

The Writeable implementation that is returned by ListFormatter::format. See the writeable crate for how to consume this.

ListFormatter

A formatter that renders sequences of items in an i18n-friendly way. See the crate-level documentation for more details.

ListFormatterPreferences

The preferences for list formatting.