icu::list::provider

Struct ListFormatterPatterns

pub struct ListFormatterPatterns<'data> {
    pub start: ListJoinerPattern<'data>,
    pub middle: VarZeroCow<'data, str>,
    pub end: ConditionalListJoinerPattern<'data>,
    pub pair: Option<ConditionalListJoinerPattern<'data>>,
}

Symbols and metadata required for ListFormatter.

๐ง This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Fields
start: ListJoinerPattern<'data>

The start pattern

middle: VarZeroCow<'data, str>

The middle pattern. It doesnโt need to be a pattern because it has to start with {0} and end with {1}, so we just store the string in between.

end: ConditionalListJoinerPattern<'data>

The end pattern

pair: Option<ConditionalListJoinerPattern<'data>>

The pair pattern, if itโs different from the end pattern.

Implementations

impl ListFormatterPatterns<'_>

pub const NARROW: &'static DataMarkerAttributes

The marker attributes for narrow lists

pub const SHORT: &'static DataMarkerAttributes

The marker attributes for short lists

pub const WIDE: &'static DataMarkerAttributes

The marker attributes for wide lists