Struct icu::list::provider::SerdeDFA

pub struct SerdeDFA<'data> { /* private fields */ }

A serde-compatible version of regex_automata::dfa::sparse::DFA. This does not implement serde::Deserialize directly, as binary deserialization is not supported in big-endian platforms. Self::maybe_deserialize can be used to deserialize to Option<SerdeDFA>.

๐ง This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Implementations

impl<'data> SerdeDFA<'data>

pub fn maybe_deserialize<'de, D>( deserializer: D ) -> Result<Option<SerdeDFA<'data>>, <D as Deserializer<'de>>::Error>where 'de: 'data, D: Deserializer<'de>,

Deserializes to Option<Self>. Will return None for non-human-readable serialization formats on big-endian systems, as regex_automata serialization is endian-sensitive.

impl<'data> SerdeDFA<'data>

pub const unsafe fn from_dfa_bytes_unchecked( dfa_bytes: &'data [u8] ) -> SerdeDFA<'data>

Creates a SerdeDFA from raw bytes. Used internally by databake.

Safety

dfa_bytes has to be a valid DFA (regex_automata::dfa::sparse::DFA::from_bytes(dfa_bytes).is_ok())

pub fn deref(&'data self) -> DFA<&'data [u8]>

Returns the represented DFA