icu::locale::extensions::unicode

Struct Unicode

pub struct Unicode {
    pub keywords: Keywords,
    pub attributes: Attributes,
}

Unicode Extensions provide information about user preferences in a given locale.

A list of Unicode BCP47 U Extensions as defined in Unicode Locale Identifier specification.

Unicode extensions provide subtags that specify language and/or locale-based behavior or refinements to language tags, according to work done by the Unicode Consortium. (See RFC 6067 for details).

Examples

use icu::locale::extensions::unicode::{key, value};
use icu::locale::Locale;

let loc: Locale =
    "de-u-hc-h12-ca-buddhist".parse().expect("Parsing failed.");

assert_eq!(
    loc.extensions.unicode.keywords.get(&key!("ca")),
    Some(&value!("buddhist"))
);

Fields
keywords: Keywords

The key-value pairs present in this locale extension, with each extension key subtag associated to its provided value subtag.

attributes: Attributes

A canonically ordered sequence of single standalone subtags for this locale extension.

Implementations

impl Unicode

pub const fn new() -> Unicode

Returns a new empty map of Unicode extensions. Same as default(), but is const.

Examples

use icu::locale::extensions::unicode::Unicode;

assert_eq!(Unicode::new(), Unicode::default());

pub fn try_from_str(s: &str) -> Result<Unicode, ParseError>

A constructor which takes a str slice, parses it and produces a well-formed Unicode.

โจ Enabled with the alloc Cargo feature.

pub fn try_from_utf8(code_units: &[u8]) -> Result<Unicode, ParseError>

See Self::try_from_str

โจ Enabled with the alloc Cargo feature.

pub fn is_empty(&self) -> bool

Returns true if there list of keywords and attributes is empty.

Examples

use icu::locale::Locale;

let loc: Locale = "en-US-u-foo".parse().expect("Parsing failed.");

assert!(!loc.extensions.unicode.is_empty());

pub fn clear(&mut self)

Clears all Unicode extension keywords and attributes, effectively removing the Unicode extension.

Example

use icu::locale::Locale;

let mut loc: Locale =
    "und-t-mul-u-hello-ca-buddhist-hc-h12".parse().unwrap();
loc.extensions.unicode.clear();
assert_eq!(loc, "und-t-mul".parse().unwrap());

pub fn total_cmp(&self, other: &Unicode) -> Ordering

Returns an ordering suitable for use in BTreeSet.

The ordering may or may not be equivalent to string ordering, and it may or may not be stable across ICU4X releases.

pub fn extend(&mut self, other: Unicode)

Extends the Unicode with values from another Unicode.

Example

use icu::locale::extensions::unicode::Unicode;

let mut ue: Unicode = "u-foobar-ca-buddhist".parse().unwrap();
let ue2: Unicode = "u-ca-gregory-hc-h12".parse().unwrap();

ue.extend(ue2);

assert_eq!(ue, "u-foobar-ca-gregory-hc-h12".parse().unwrap());

impl Unicode

pub fn to_string(&self) -> String

Converts the given value to a String.

Under the hood, this uses an efficient [Writeable] implementation. However, in order to avoid allocating a string, it is more efficient to use [Writeable] directly.