icu::locale

Module provider

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Structs

Aliases

This alias data is used for locale canonicalization.

Baked

Baked data

ExemplarCharactersData

A set of characters and strings which share a particular property value.

LanguageStrStrPair

A triplet of strings with a EncodeAsVarULE implementation.

LanguageStrStrPairVarULE

VarULE type for LanguageStrStrPair. See LanguageStrStrPair for documentation.

LikelySubtagsExtended

This likely subtags data is used for full coverage of locales, including ones that don’t otherwise have data in the Common Locale Data Repository (CLDR).

LikelySubtagsForLanguage

This likely subtags data is used for the minimize and maximize operations.

LikelySubtagsForScriptRegion

This likely subtags data is used for the minimize and maximize operations.

LocaleAliasesV1

Marker for locale alias data.

LocaleExemplarCharactersAuxiliaryV1

Marker for auxiliary exemplar characters data.

LocaleExemplarCharactersIndexV1

Marker for index exemplar characters data.

LocaleExemplarCharactersMainV1

Marker for main exemplar characters data.

LocaleExemplarCharactersNumbersV1

Marker for numbers exemplar characters data.

LocaleExemplarCharactersPunctuationV1

Marker for punctuation exemplar characters data.

LocaleLikelySubtagsExtendedV1

Marker for extended data for likely subtags.

LocaleLikelySubtagsLanguageV1

Marker for data for likely subtags for languages.

LocaleLikelySubtagsScriptRegionV1

Marker for data for likely subtags for scripts and regions.

LocaleParentsV1

Marker for locale fallback parents data.

LocaleScriptDirectionV1

Marker for script direction data.

Parents

Locale fallback rules derived from CLDR parent locales data.

ScriptDirection

This directionality data is used to determine the script directionality of a locale.

StrStrPair

A pair of strings with a EncodeAsVarULE implementation.

StrStrPairVarULE

VarULE type for StrStrPair. See StrStrPair for documentation.