Module icu::locid

Modules

extensions
Unicode Extensions provide a mechanism to extend the LanguageIdentifier with additional bits of information - a combination of a LanguageIdentifier and Extensions is called Locale.

macros

subtags
Language Identifier and Locale contains a set of subtags which represent different fields of the structure.

Structs

LanguageIdentifier
LanguageIdentifier is a core struct representing a Unicode BCP47 Language Identifier.

Locale
Locale is a core struct representing a Unicode Locale Identifier.

Enums

ParserError
List of parser errors that can be generated while parsing LanguageIdentifier, Locale, subtags or extensions.