Module icu::locid::extensions::transform

Transform Extensions provide information on content transformations in a given locale.

The main struct for this extension is Transform which contains Fields and an optional LanguageIdentifier.

Examples

use icu_locid::{LanguageIdentifier, Locale};
use icu_locid::extensions::transform::{Transform, Fields, Key, Value};

let mut loc: Locale = "en-US-t-es-AR-h0-hybrid".parse()
    .expect("Parsing failed.");

let lang: LanguageIdentifier = "es-AR".parse()
    .expect("Parsing LanguageIdentifier failed.");

let key: Key = "h0".parse()
    .expect("Parsing key failed.");
let value: Value = "hybrid".parse()
    .expect("Parsing value failed.");

assert_eq!(loc.extensions.transform.lang, Some(lang));
assert!(loc.extensions.transform.fields.contains_key(key));
assert_eq!(loc.extensions.transform.fields.get(key), Some(&value));

assert_eq!(&loc.extensions.transform.to_string(), "-t-es-AR-h0-hybrid");

Structs

Fields
A list of Key-Value pairs representing functional information about content transformations.

Key
A key used in a list of Fields.

Transform
A list of Unicode BCP47 T Extensions as defined in Unicode Locale Identifier specification.

Value