Struct icu::locid::extensions::unicode::Key

pub struct Key(_);

A key used in a list of Keywords.

The key has to be a two ASCII alphanumerical characters long, with the first character being alphanumeric, and the second being alphabetic.

Examples

use icu::locid::extensions::unicode::Key;

let key1: Key = "ca".parse()
    .expect("Failed to parse a Key.");

assert_eq!(key1, "ca");

Implementations

impl Key

pub fn from_bytes(key: &[u8]) -> Result<Key, ParserError>

A constructor which takes a utf8 slice, parses it and produces a well-formed Key.

Examples

use icu::locid::extensions::unicode::Key;

let key = Key::from_bytes(b"ca")
    .expect("Parsing failed.");

assert_eq!(key, "ca");

pub fn from_tinystr4_unchecked(key: TinyStr4) -> Key

A constructor which takes a TinyStr4 and produces a Key without doing any checking.

Examples

use icu::locid::extensions::unicode::Key;
use tinystr::tinystr4;

let key = Key::from_tinystr4_unchecked(tinystr4!("ca"));

assert_eq!(key, "ca");

pub fn as_str(&self) -> &str

A helper function for displaying a Key subtag as a &str.

Examples

use icu::locid::extensions::unicode::Key;

let key: Key = "ca".parse().expect("Parsing failed.");

assert_eq!(key.as_str(), "ca");

Notice: For many use cases, such as comparison, Key implements PartialEq<&str> which allows for direct comparisons.