Enum icu::properties::GeneralCategory

#[repr(u8)]
pub enum GeneralCategory {

    Unassigned,
    UppercaseLetter,
    LowercaseLetter,
    TitlecaseLetter,
    ModifierLetter,
    OtherLetter,
    NonspacingMark,
    SpacingMark,
    EnclosingMark,
    DecimalNumber,
    LetterNumber,
    OtherNumber,
    SpaceSeparator,
    LineSeparator,
    ParagraphSeparator,
    Control,
    Format,
    PrivateUse,
    Surrogate,
    DashPunctuation,
    OpenPunctuation,
    ClosePunctuation,
    ConnectorPunctuation,
    InitialPunctuation,
    FinalPunctuation,
    OtherPunctuation,
    MathSymbol,
    CurrencySymbol,
    ModifierSymbol,
    OtherSymbol,

}

Enumerated property General_Category.

General_Category specifies the most general classification of a code point, usually determined based on the primary characteristic of the assigned character. For example, is the character a letter, a mark, a number, punctuation, or a symbol, and if so, of what type?

GeneralCategory only supports specific subcategories (eg UppercaseLetter). It does not support grouped categories (eg Letter). For grouped categories, use GeneralCategoryGroup.

Variants

Unassigned

(Cn) A reserved unassigned code point or a noncharacter

UppercaseLetter

(Lu) An uppercase letter

LowercaseLetter

(Ll) A lowercase letter

TitlecaseLetter

(Lt) A digraphic letter, with first part uppercase

ModifierLetter

(Lm) A modifier letter

OtherLetter

(Lo) Other letters, including syllables and ideographs

NonspacingMark

(Mn) A nonspacing combining mark (zero advance width)

SpacingMark

(Mc) A spacing combining mark (positive advance width)

EnclosingMark

(Me) An enclosing combining mark

DecimalNumber

(Nd) A decimal digit

LetterNumber

(Nl) A letterlike numeric character

OtherNumber

(No) A numeric character of other type

SpaceSeparator

(Zs) A space character (of various non-zero widths)

LineSeparator

(Zl) U+2028 LINE SEPARATOR only

ParagraphSeparator

(Zp) U+2029 PARAGRAPH SEPARATOR only

Control

(Cc) A C0 or C1 control code

Format

(Cf) A format control character

PrivateUse

(Co) A private-use character

Surrogate

(Cs) A surrogate code point

DashPunctuation

(Pd) A dash or hyphen punctuation mark

OpenPunctuation

(Ps) An opening punctuation mark (of a pair)

ClosePunctuation

(Pe) A closing punctuation mark (of a pair)

ConnectorPunctuation

(Pc) A connecting punctuation mark, like a tie

InitialPunctuation

(Pi) An initial quotation mark

FinalPunctuation

(Pf) A final quotation mark

OtherPunctuation

(Po) A punctuation mark of other type

MathSymbol

(Sm) A symbol of mathematical use

CurrencySymbol

(Sc) A currency sign

ModifierSymbol

(Sk) A non-letterlike modifier symbol

OtherSymbol

(So) A symbol of other type