Module icu::properties::maps

The functions in this module return a CodePointMapData representing, for each code point in the entire range of code points, the property values for a particular Unicode property.

The descriptions of most properties are taken from TR44, the documentation for the Unicode Character Database.

Structs

CodePointMapData

A wrapper around code point map data. It is returned by APIs that return Unicode property data in a map-like form, ex: enumerated property value data keyed by code point. Access its data via the borrowed version, CodePointMapDataBorrowed.

CodePointMapDataBorrowed

A borrowed wrapper around code point set data, returned by CodePointSetData::as_borrowed(). More efficient to query.

Functions

bidi_class

Return a CodePointMapDataBorrowed for the Bidi_Class Unicode enumerated property. See [BidiClass].

canonical_combining_class

Return a CodePointMapData for the Canonical_Combining_Class Unicode property. See [CanonicalCombiningClass].

east_asian_width

Return a CodePointMapDataBorrowed for the East_Asian_Width Unicode enumerated property. See [EastAsianWidth].

general_category

Return a CodePointMapDataBorrowed for the General_Category Unicode enumerated property. See [GeneralCategory].

grapheme_cluster_break

Return a CodePointMapDataBorrowed for the Grapheme_Cluster_Break Unicode enumerated property. See [GraphemeClusterBreak].

indic_syllabic_category

Return a CodePointMapData for the Indic_Syllabic_Category Unicode property. See [IndicSyllabicCategory].

line_break

Return a CodePointMapDataBorrowed for the Line_Break Unicode enumerated property. See [LineBreak].

load_bidi_class

A version of bidi_class() that uses custom data provided by a DataProvider.

load_canonical_combining_class

A version of canonical_combining_class() that uses custom data provided by a DataProvider.

load_east_asian_width

A version of east_asian_width() that uses custom data provided by a DataProvider.

load_general_category

A version of general_category() that uses custom data provided by a DataProvider.

load_grapheme_cluster_break

A version of grapheme_cluster_break() that uses custom data provided by a DataProvider.

load_indic_syllabic_category

A version of indic_syllabic_category() that uses custom data provided by a DataProvider.

load_line_break

A version of line_break() that uses custom data provided by a DataProvider.

load_script

A version of script() that uses custom data provided by a DataProvider.

load_sentence_break

A version of sentence_break() that uses custom data provided by a DataProvider.

load_word_break

A version of word_break() that uses custom data provided by a DataProvider.

script

Return a CodePointMapDataBorrowed for the Script Unicode enumerated property. See [Script].

sentence_break

Return a CodePointMapDataBorrowed for the Sentence_Break Unicode enumerated property. See [SentenceBreak].

word_break

Return a CodePointMapDataBorrowed for the Word_Break Unicode enumerated property. See [WordBreak].