Module icu::properties::names

Module for working with the names of property values

Structs

PropertyEnumToValueNameLinearMapper

A struct capable of looking up a property name from a value Access its data by calling Self::as_borrowed() and using the methods on PropertyEnumToValueNameLinearMapperBorrowed.

PropertyEnumToValueNameLinearMapperBorrowed

A borrowed wrapper around property value name-to-enum data, returned by PropertyEnumToValueNameLinearMapper::as_borrowed(). More efficient to query.

PropertyEnumToValueNameLinearTiny4Mapper

A struct capable of looking up a property name from a value Access its data by calling Self::as_borrowed() and using the methods on PropertyEnumToValueNameLinearTiny4MapperBorrowed.

PropertyEnumToValueNameLinearTiny4MapperBorrowed

A borrowed wrapper around property value name-to-enum data, returned by PropertyEnumToValueNameLinearTiny4Mapper::as_borrowed(). More efficient to query.

PropertyEnumToValueNameSparseMapper

A struct capable of looking up a property name from a value Access its data by calling Self::as_borrowed() and using the methods on PropertyEnumToValueNameSparseMapperBorrowed.

PropertyEnumToValueNameSparseMapperBorrowed

A borrowed wrapper around property value name-to-enum data, returned by PropertyEnumToValueNameSparseMapper::as_borrowed(). More efficient to query.

PropertyValueNameToEnumMapper

A struct capable of looking up a property value from a string name. Access its data by calling Self::as_borrowed() and using the methods on PropertyValueNameToEnumMapperBorrowed.

PropertyValueNameToEnumMapperBorrowed

A borrowed wrapper around property value name-to-enum data, returned by PropertyValueNameToEnumMapper::as_borrowed(). More efficient to query.