icu::properties

Module provider

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Modules

names

🚧 [Unstable] Property names-related data for this component

Structs

Baked

Baked data

GeneralCategoryULE

ULE type for GeneralCategory

PropertyBinaryAlnumV1

PropertyBinaryAlnumV1

PropertyBinaryAlphabeticV1

PropertyBinaryAlphabeticV1

PropertyBinaryAsciiHexDigitV1

PropertyBinaryAsciiHexDigitV1

PropertyBinaryBasicEmojiV1

PropertyBinaryBasicEmojiV1

PropertyBinaryBidiControlV1

PropertyBinaryBidiControlV1

PropertyBinaryBidiMirroredV1

PropertyBinaryBidiMirroredV1

PropertyBinaryBlankV1

PropertyBinaryBlankV1

PropertyBinaryCaseIgnorableV1

PropertyBinaryCaseIgnorableV1

PropertyBinaryCaseSensitiveV1

PropertyBinaryCaseSensitiveV1

PropertyBinaryCasedV1

PropertyBinaryCasedV1

PropertyBinaryChangesWhenCasefoldedV1

PropertyBinaryChangesWhenCasefoldedV1

PropertyBinaryChangesWhenCasemappedV1

PropertyBinaryChangesWhenCasemappedV1

PropertyBinaryChangesWhenLowercasedV1

PropertyBinaryChangesWhenLowercasedV1

PropertyBinaryChangesWhenNfkcCasefoldedV1

PropertyBinaryChangesWhenNfkcCasefoldedV1

PropertyBinaryChangesWhenTitlecasedV1

PropertyBinaryChangesWhenTitlecasedV1

PropertyBinaryChangesWhenUppercasedV1

PropertyBinaryChangesWhenUppercasedV1

PropertyBinaryDashV1

PropertyBinaryDashV1

PropertyBinaryDefaultIgnorableCodePointV1

PropertyBinaryDefaultIgnorableCodePointV1

PropertyBinaryDeprecatedV1

PropertyBinaryDeprecatedV1

PropertyBinaryDiacriticV1

PropertyBinaryDiacriticV1

PropertyBinaryEmojiComponentV1

PropertyBinaryEmojiComponentV1

PropertyBinaryEmojiModifierBaseV1

PropertyBinaryEmojiModifierBaseV1

PropertyBinaryEmojiModifierV1

PropertyBinaryEmojiModifierV1

PropertyBinaryEmojiPresentationV1

PropertyBinaryEmojiPresentationV1

PropertyBinaryEmojiV1

PropertyBinaryEmojiV1

PropertyBinaryExtendedPictographicV1

PropertyBinaryExtendedPictographicV1

PropertyBinaryExtenderV1

PropertyBinaryExtenderV1

PropertyBinaryFullCompositionExclusionV1

PropertyBinaryFullCompositionExclusionV1

PropertyBinaryGraphV1

PropertyBinaryGraphV1

PropertyBinaryGraphemeBaseV1

PropertyBinaryGraphemeBaseV1

PropertyBinaryGraphemeExtendV1

PropertyBinaryGraphemeExtendV1

PropertyBinaryGraphemeLinkV1

PropertyBinaryGraphemeLinkV1

PropertyBinaryHexDigitV1

PropertyBinaryHexDigitV1

PropertyBinaryHyphenV1

PropertyBinaryHyphenV1

PropertyBinaryIdCompatMathContinueV1

PropertyBinaryIdCompatMathContinueV1

PropertyBinaryIdCompatMathStartV1

PropertyBinaryIdCompatMathStartV1

PropertyBinaryIdContinueV1

PropertyBinaryIdContinueV1

PropertyBinaryIdStartV1

PropertyBinaryIdStartV1

PropertyBinaryIdeographicV1

PropertyBinaryIdeographicV1

PropertyBinaryIdsBinaryOperatorV1

PropertyBinaryIdsBinaryOperatorV1

PropertyBinaryIdsTrinaryOperatorV1

PropertyBinaryIdsTrinaryOperatorV1

PropertyBinaryIdsUnaryOperatorV1

PropertyBinaryIdsUnaryOperatorV1

PropertyBinaryJoinControlV1

PropertyBinaryJoinControlV1

PropertyBinaryLogicalOrderExceptionV1

PropertyBinaryLogicalOrderExceptionV1

PropertyBinaryLowercaseV1

PropertyBinaryLowercaseV1

PropertyBinaryMathV1

PropertyBinaryMathV1

PropertyBinaryModifierCombiningMarkV1

PropertyBinaryModifierCombiningMarkV1

PropertyBinaryNfcInertV1

PropertyBinaryNfcInertV1

PropertyBinaryNfdInertV1

PropertyBinaryNfdInertV1

PropertyBinaryNfkcInertV1

PropertyBinaryNfkcInertV1

PropertyBinaryNfkdInertV1

PropertyBinaryNfkdInertV1

PropertyBinaryNoncharacterCodePointV1

PropertyBinaryNoncharacterCodePointV1

PropertyBinaryPatternSyntaxV1

PropertyBinaryPatternSyntaxV1

PropertyBinaryPatternWhiteSpaceV1

PropertyBinaryPatternWhiteSpaceV1

PropertyBinaryPrependedConcatenationMarkV1

PropertyBinaryPrependedConcatenationMarkV1

PropertyBinaryPrintV1

PropertyBinaryPrintV1

PropertyBinaryQuotationMarkV1

PropertyBinaryQuotationMarkV1

PropertyBinaryRadicalV1

PropertyBinaryRadicalV1

PropertyBinaryRegionalIndicatorV1

PropertyBinaryRegionalIndicatorV1

PropertyBinarySegmentStarterV1

PropertyBinarySegmentStarterV1

PropertyBinarySentenceTerminalV1

PropertyBinarySentenceTerminalV1

PropertyBinarySoftDottedV1

PropertyBinarySoftDottedV1

PropertyBinaryTerminalPunctuationV1

PropertyBinaryTerminalPunctuationV1

PropertyBinaryUnifiedIdeographV1

PropertyBinaryUnifiedIdeographV1

PropertyBinaryUppercaseV1

PropertyBinaryUppercaseV1

PropertyBinaryVariationSelectorV1

PropertyBinaryVariationSelectorV1

PropertyBinaryWhiteSpaceV1

PropertyBinaryWhiteSpaceV1

PropertyBinaryXdigitV1

PropertyBinaryXdigitV1

PropertyBinaryXidContinueV1

PropertyBinaryXidContinueV1

PropertyBinaryXidStartV1

PropertyBinaryXidStartV1

PropertyEnumBidiClassV1

Data marker for the ‘BidiClass’ Unicode property

PropertyEnumBidiMirroringGlyphV1

Data marker for the ‘BidiMirroringGlyph’ Unicode property

PropertyEnumCanonicalCombiningClassV1

Data marker for the ‘CanonicalCombiningClass’ Unicode property

PropertyEnumEastAsianWidthV1

Data marker for the ‘EastAsianWidth’ Unicode property

PropertyEnumGeneralCategoryV1

Data marker for the ‘GeneralCategory’ Unicode property

PropertyEnumGraphemeClusterBreakV1

Data marker for the ‘GraphemeClusterBreak’ Unicode property

PropertyEnumHangulSyllableTypeV1

Data marker for the ‘HangulSyllableType’ Unicode property

PropertyEnumIndicConjunctBreakV1

Data marker for the ‘IndicConjunctBreak’ Unicode property

PropertyEnumIndicSyllabicCategoryV1

Data marker for the ‘IndicSyllabicCategory’ Unicode property

PropertyEnumJoiningTypeV1

Data marker for the ‘JoiningType’ Unicode property

PropertyEnumLineBreakV1

Data marker for the ‘LineBreak’ Unicode property

PropertyEnumScriptV1

Data marker for the ‘Script’ Unicode property

PropertyEnumSentenceBreakV1

Data marker for the ‘SentenceBreak’ Unicode property

PropertyEnumVerticalOrientationV1

Data marker for the ‘Vertical_Orientation’ Unicode property

PropertyEnumWordBreakV1

Data marker for the ‘WordBreak’ Unicode property

PropertyNameLongBidiClassV1

PropertyNameLongBidiClassV1

PropertyNameLongCanonicalCombiningClassV1

PropertyNameLongCanonicalCombiningClassV1

PropertyNameLongEastAsianWidthV1

PropertyNameLongEastAsianWidthV1

PropertyNameLongGeneralCategoryV1

PropertyNameLongGeneralCategoryV1

PropertyNameLongGraphemeClusterBreakV1

PropertyNameLongGraphemeClusterBreakV1

PropertyNameLongHangulSyllableTypeV1

PropertyNameLongHangulSyllableTypeV1

PropertyNameLongIndicConjunctBreakV1

PropertyNameLongIndicConjunctBreakV1

PropertyNameLongIndicSyllabicCategoryV1

PropertyNameLongIndicSyllabicCategoryV1

PropertyNameLongJoiningTypeV1

PropertyNameLongJoiningTypeV1

PropertyNameLongLineBreakV1

PropertyNameLongLineBreakV1

PropertyNameLongScriptV1

PropertyNameLongScriptV1

PropertyNameLongSentenceBreakV1

PropertyNameLongSentenceBreakV1

PropertyNameLongVerticalOrientationV1

PropertyNameLongVerticalOrientationV1

PropertyNameLongWordBreakV1

PropertyNameLongWordBreakV1

PropertyNameParseBidiClassV1

PropertyNameParseBidiClassV1

PropertyNameParseCanonicalCombiningClassV1

PropertyNameParseCanonicalCombiningClassV1

PropertyNameParseEastAsianWidthV1

PropertyNameParseEastAsianWidthV1

PropertyNameParseGeneralCategoryMaskV1

PropertyNameParseGeneralCategoryMaskV1

PropertyNameParseGeneralCategoryV1

PropertyNameParseGeneralCategoryV1

PropertyNameParseGraphemeClusterBreakV1

PropertyNameParseGraphemeClusterBreakV1

PropertyNameParseHangulSyllableTypeV1

PropertyNameParseHangulSyllableTypeV1

PropertyNameParseIndicConjunctBreakV1

PropertyNameParseIndicConjunctBreakV1

PropertyNameParseIndicSyllabicCategoryV1

PropertyNameParseIndicSyllabicCategoryV1

PropertyNameParseJoiningTypeV1

PropertyNameParseJoiningTypeV1

PropertyNameParseLineBreakV1

PropertyNameParseLineBreakV1

PropertyNameParseScriptV1

PropertyNameParseScriptV1

PropertyNameParseSentenceBreakV1

PropertyNameParseSentenceBreakV1

PropertyNameParseVerticalOrientationV1

PropertyNameParseVerticalOrientationV1

PropertyNameParseWordBreakV1

PropertyNameParseWordBreakV1

PropertyNameShortBidiClassV1

PropertyNameShortBidiClassV1

PropertyNameShortCanonicalCombiningClassV1

PropertyNameShortCanonicalCombiningClassV1

PropertyNameShortEastAsianWidthV1

PropertyNameShortEastAsianWidthV1

PropertyNameShortGeneralCategoryV1

PropertyNameShortGeneralCategoryV1

PropertyNameShortGraphemeClusterBreakV1

PropertyNameShortGraphemeClusterBreakV1

PropertyNameShortHangulSyllableTypeV1

PropertyNameShortHangulSyllableTypeV1

PropertyNameShortIndicConjunctBreakV1

PropertyNameShortIndicConjunctBreakV1

PropertyNameShortIndicSyllabicCategoryV1

PropertyNameShortIndicSyllabicCategoryV1

PropertyNameShortJoiningTypeV1

PropertyNameShortJoiningTypeV1

PropertyNameShortLineBreakV1

PropertyNameShortLineBreakV1

PropertyNameShortScriptV1

PropertyNameShortScriptV1

PropertyNameShortSentenceBreakV1

PropertyNameShortSentenceBreakV1

PropertyNameShortVerticalOrientationV1

PropertyNameShortVerticalOrientationV1

PropertyNameShortWordBreakV1

PropertyNameShortWordBreakV1

PropertyScriptWithExtensionsV1

PropertyScriptWithExtensionsV1

ScriptWithExtensionsProperty

A struct that efficiently stores Script and Script_Extensions property data.

Enums

PropertyCodePointMap

A map efficiently storing data about individual characters.

PropertyCodePointSet

A set of characters which share a particular property value.

PropertyUnicodeSet

A set of characters and strings which share a particular property value.

Constants

MARKERS

All data keys in this module.