Enum icu::properties::provider::bidi_data::CheckedBidiPairedBracketType

#[repr(u8)]
pub enum CheckedBidiPairedBracketType {
    None,
    Open,
    Close,
}

A closed Rust enum representing a closed set of the incoming Bidi_Paired_Bracket_Type property values necessary in the internal representation of MirroredPairedBracketData to satisfy the ULE invariants on valid values.

Variants

None

Not a paired bracket.

Open

Open paired bracket.

Close

Close paired bracket.