Struct icu::properties::provider::names::NormalizedPropertyNameStr

#[repr(transparent)]
pub struct NormalizedPropertyNameStr(_);

This is a property name that can be โloose matchedโ as according to PropertyValueAliases.txt

(matched case-insensitively in ASCII, ignoring underscores, whitespace, and hyphens)

This is expected to be ASCII, but we do not rely on this invariant anywhere except during datagen.

The Ord impl will sort things using strict equality, but in such a way that all loose-equal items will sort into the same area, such that a map can be searched for both strict and loose equality.

๐ง This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Implementations

impl NormalizedPropertyNameStr

pub fn boxed_from_bytes(b: &[u8]) -> Box<NormalizedPropertyNameStr, Global>

Get a Box<NormalizedPropertyName> from a byte slice