Struct icu::properties::provider::ScriptWithExtensionsPropertyV1

pub struct ScriptWithExtensionsPropertyV1<'data> {
    pub trie: CodePointTrie<'data, ScriptWithExt>,
    pub extensions: VarZeroVec<'data, ZeroSlice<Script>, Index16>,
}

A struct that efficiently stores Script and Script_Extensions property data.

๐ง This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Fields
trie: CodePointTrie<'data, ScriptWithExt>

Note: The ScriptWithExt values in this array will assume a 12-bit layout. The 2 higher order bits 11..10 will indicate how to deduce the Script value and Script_Extensions value, nearly matching the representation in ICU:

High order 2 bits value	Script	Script_Extensions

3	First value in sub-array, index given by lower 10 bits	Sub-array excluding first value, index given by lower 10 bits

2	Script=Inherited	Entire sub-array, index given by lower 10 bits

1	Script=Common	Entire sub-array, index given by lower 10 bits

0	Value in lower 10 bits	[ Script value ] single-element array

When the lower 10 bits of the value are used as an index, that index is used for the outer-level vector of the nested extensions structure.

extensions: VarZeroVec<'data, ZeroSlice<Script>, Index16>

This companion structure stores Script_Extensions values, which are themselves arrays / vectors. This structure only stores the values for cases in which scx(cp) != [ sc(cp) ]. Each sub-vector is distinct. The sub-vector represents the Script_Extensions array value for a code point, and may also indicate Script value, as described for the trie field.