Struct icu::properties::GeneralCategoryGroup

#[repr(transparent)]
pub struct GeneralCategoryGroup(_);

Groupings of multiple General_Category property values.

Instances of GeneralCategoryGroup represent the defined multi-category values that are useful for users in certain contexts, such as regex. In other words, unlike GeneralCategory, this supports groups of general categories: for example, Letter /// is the union of UppercaseLetter, LowercaseLetter, etc.

See https://www.unicode.org/reports/tr44/ .

The discriminants correspond to the U_GC_XX_MASK constants in ICU4C. Unlike GeneralCategory, this supports groups of general categories: for example, Letter is the union of UppercaseLetter, LowercaseLetter, etc.

See UCharCategory and U_GET_GC_MASK in ICU4C.

Implementations

impl GeneralCategoryGroup

pub const UppercaseLetter: GeneralCategoryGroup

(Lu) An uppercase letter

pub const LowercaseLetter: GeneralCategoryGroup

(Ll) A lowercase letter

pub const TitlecaseLetter: GeneralCategoryGroup

(Lt) A digraphic letter, with first part uppercase

pub const ModifierLetter: GeneralCategoryGroup

(Lm) A modifier letter

pub const OtherLetter: GeneralCategoryGroup

(Lo) Other letters, including syllables and ideographs

pub const CasedLetter: GeneralCategoryGroup

(LC) The union of UppercaseLetter, LowercaseLetter, and TitlecaseLetter

pub const Letter: GeneralCategoryGroup

(L) The union of all letter categories

pub const NonspacingMark: GeneralCategoryGroup

(Mn) A nonspacing combining mark (zero advance width)

pub const EnclosingMark: GeneralCategoryGroup

(Mc) A spacing combining mark (positive advance width)

pub const SpacingMark: GeneralCategoryGroup

(Me) An enclosing combining mark

pub const Mark: GeneralCategoryGroup

(M) The union of all mark categories

pub const DecimalNumber: GeneralCategoryGroup

(Nd) A decimal digit

pub const LetterNumber: GeneralCategoryGroup

(Nl) A letterlike numeric character

pub const OtherNumber: GeneralCategoryGroup

(No) A numeric character of other type

pub const Number: GeneralCategoryGroup

(N) The union of all number categories

pub const SpaceSeparator: GeneralCategoryGroup

(Zs) A space character (of various non-zero widths)

pub const LineSeparator: GeneralCategoryGroup

(Zl) U+2028 LINE SEPARATOR only

pub const ParagraphSeparator: GeneralCategoryGroup

(Zp) U+2029 PARAGRAPH SEPARATOR only

pub const Separator: GeneralCategoryGroup

(Z) The union of all separator categories

pub const Control: GeneralCategoryGroup

(Cc) A C0 or C1 control code

pub const Format: GeneralCategoryGroup

(Cf) A format control character

pub const PrivateUse: GeneralCategoryGroup

(Co) A private-use character

pub const Surrogate: GeneralCategoryGroup

(Cs) A surrogate code point

pub const Unassigned: GeneralCategoryGroup

(Cn) A reserved unassigned code point or a noncharacter

pub const Other: GeneralCategoryGroup

(C) The union of all control code, reserved, and unassigned categories

pub const DashPunctuation: GeneralCategoryGroup

(Pd) A dash or hyphen punctuation mark

pub const OpenPunctuation: GeneralCategoryGroup

(Ps) An opening punctuation mark (of a pair)

pub const ClosePunctuation: GeneralCategoryGroup

(Pe) A closing punctuation mark (of a pair)

pub const ConnectorPunctuation: GeneralCategoryGroup

(Pc) A connecting punctuation mark, like a tie

pub const InitialPunctuation: GeneralCategoryGroup

(Pi) An initial quotation mark

pub const FinalPunctuation: GeneralCategoryGroup

(Pf) A final quotation mark

pub const OtherPunctuation: GeneralCategoryGroup

(Po) A punctuation mark of other type

pub const Punctuation: GeneralCategoryGroup

(P) The union of all punctuation categories

pub const MathSymbol: GeneralCategoryGroup

(Sm) A symbol of mathematical use

pub const CurrencySymbol: GeneralCategoryGroup

(Sc) A currency sign

pub const ModifierSymbol: GeneralCategoryGroup

(Sk) A non-letterlike modifier symbol

pub const OtherSymbol: GeneralCategoryGroup

(So) A symbol of other type

pub const Symbol: GeneralCategoryGroup

(S) The union of all symbol categories

pub fn contains(&self, val: GeneralCategory) -> bool

Return whether the code point belongs in the provided multi-value category.

use icu::properties::{maps, GeneralCategory, GeneralCategoryGroup};
use icu_codepointtrie::CodePointTrie;

let provider = icu_testdata::get_provider();
let payload =
    maps::get_general_category(&provider)
        .expect("The data should be valid");
let data_struct = payload.get();
let gc = &data_struct.code_point_trie;

assert_eq!(gc.get('A' as u32), GeneralCategory::UppercaseLetter);
assert!(GeneralCategoryGroup::CasedLetter.contains(gc.get('A' as u32)));

// U+0B1E ORIYA LETTER NYA
assert_eq!(gc.get('ଞ' as u32), GeneralCategory::OtherLetter);
assert!(GeneralCategoryGroup::Letter.contains(gc.get('ଞ' as u32)));
assert!(!GeneralCategoryGroup::CasedLetter.contains(gc.get('ଞ' as u32)));

// U+0301 COMBINING ACUTE ACCENT
assert_eq!(gc.get(0x0301), GeneralCategory::NonspacingMark);
assert!(GeneralCategoryGroup::Mark.contains(gc.get(0x0301)));
assert!(!GeneralCategoryGroup::Letter.contains(gc.get(0x0301)));

assert_eq!(gc.get('0' as u32), GeneralCategory::DecimalNumber);
assert!(GeneralCategoryGroup::Number.contains(gc.get('0' as u32)));
assert!(!GeneralCategoryGroup::Mark.contains(gc.get('0' as u32)));

assert_eq!(gc.get('(' as u32), GeneralCategory::OpenPunctuation);
assert!(GeneralCategoryGroup::Punctuation.contains(gc.get('(' as u32)));
assert!(!GeneralCategoryGroup::Number.contains(gc.get('(' as u32)));

// U+2713 CHECK MARK
assert_eq!(gc.get('✓' as u32), GeneralCategory::OtherSymbol);
assert!(GeneralCategoryGroup::Symbol.contains(gc.get('✓' as u32)));
assert!(!GeneralCategoryGroup::Punctuation.contains(gc.get('✓' as u32)));

assert_eq!(gc.get(' ' as u32), GeneralCategory::SpaceSeparator);
assert!(GeneralCategoryGroup::Separator.contains(gc.get(' ' as u32)));
assert!(!GeneralCategoryGroup::Symbol.contains(gc.get(' ' as u32)));

// U+E007F CANCEL TAG
assert_eq!(gc.get(0xE007F), GeneralCategory::Format);
assert!(GeneralCategoryGroup::Other.contains(gc.get(0xE007F)));
assert!(!GeneralCategoryGroup::Separator.contains(gc.get(0xE007F)));