icu::segmenter

Module provider

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Structs

Baked

Baked data

LstmDataFloat32

The struct that stores a LSTM model.

LstmMatrix1

The struct that stores a LSTM’s matrix.

LstmMatrix2

The struct that stores a LSTM’s matrix.

LstmMatrix3

The struct that stores a LSTM’s matrix.

RuleBreakData

Pre-processed Unicode data in the form of tables to be used for rule-based breaking.

RuleBreakDataOverride

codepoint trie data that the difference by specific locale

SegmenterBreakGraphemeClusterV1

SegmenterBreakGraphemeClusterV1

SegmenterBreakLineV1

SegmenterBreakLineV1

SegmenterBreakSentenceOverrideV1

SegmenterBreakSentenceOverrideV1

SegmenterBreakSentenceV1

SegmenterBreakSentenceV1

SegmenterBreakWordOverrideV1

SegmenterBreakWordOverrideV1

SegmenterBreakWordV1

SegmenterBreakWordV1

SegmenterDictionaryAutoV1

SegmenterDictionaryWordAutoV1

SegmenterDictionaryExtendedV1

SegmenterDictionaryExtendedV1

SegmenterLstmAutoV1

SegmenterLstmWordLineAutoV1

UCharDictionaryBreakData

char16trie data for dictionary break

WordTypeULE

ULE type for WordType

Enums

BreakState

Break state

LstmData

The data to power the LSTM segmentation model.

ModelType

The type of LSTM model