icu::segmenter::scaffold

Trait RuleBreakType

pub trait RuleBreakType: Sized + Sealed {
    type IterAttr<'s>: Iterator<Item = (usize, Self::CharType)> + Clone + Debug;
    type CharType: Copy + Into<u32> + Debug;
}

A trait allowing for RuleBreakIterator to be generalized to multiple string encoding methods and granularity such as grapheme cluster, word, etc.

๐ซ This trait is sealed; it cannot be implemented by user code. If an API requests an item that implements this trait, please consider using a type from the implementors listed below.

Required Associated Types

type IterAttr<'s>: Iterator<Item = (usize, Self::CharType)> + Clone + Debug

The iterator over characters.

type CharType: Copy + Into<u32> + Debug

The character type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.