icu

Crate time

Time and timezone functionality.

This module is published as its own crate (icu_time) and as part of the icu crate. See the latter for more details on the ICU4X project.

Modules

provider

๐ง [Unstable] Data provider struct definitions for this ICU4X component.

scaffold

Scaffolding traits and types not generally required in client code.

zone

Types for resolving and manipulating time zones.

Structs

DateTime

A date and time for a given calendar.

Hour

An ISO-8601 hour component, for use with ISO calendars.

Minute

An ISO-8601 minute component, for use with ISO calendars.

Nanosecond

A fractional second component, stored as nanoseconds.

Second

An ISO-8601 second component, for use with ISO calendars.

Time

A representation of a time in hours, minutes, seconds, and nanoseconds

TimeZone

A CLDR time zone identity.

TimeZoneInfo

A utility type that can hold time zone information.

ZonedDateTime

A date and time for a given calendar, local to a specified time zone.

Enums

ParseError

The error type for parsing RFC 9557 strings.