icu::time

Module provider

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Modules

iana

🚧 [Unstable] Property names-related data for this component

windows

A provider for mapping Windows Zones to IANA identifiers.

Structs

Baked

Baked data

MetazoneInfo

Information about a metazone membership

TimeZone

A CLDR time zone identity.

TimeZoneVariantULE

ULE type for TimeZoneVariant

Timestamp24

Encodes ZoneNameTimestamp in 3 bytes by dropping the unused metadata

TimezonePeriods

Data struct for the TimezonePeriodsV1 marker.

TimezonePeriodsV1

An ICU4X mapping to timezone offset data and metazones at a given period.

VariantOffsets

Represents the different offsets in use for a time zone

VariantOffsetsWithMetazoneMembershipKind

A VariantOffsets and a MetazoneMembershipKind packed into one byte.

Enums

MetazoneMembershipKind

Metadata about a metazone membership

TimeZoneVariant

A time zone variant used to identify a display name in CLDR.

Type Aliases

MetazoneId

Metazone ID in a compact format