icu::time::provider

Module iana

🚧 [Unstable] Property names-related data for this component

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Structs

IanaNames

A mapping from IANA time zone identifiers to BCP-47 time zone identifiers.

IanaToBcp47Map

A mapping from normal-case IANA time zone identifiers to BCP-47 time zone identifiers.

TimezoneIdentifiersIanaCoreV1

See IanaToBcp47Map

TimezoneIdentifiersIanaExtendedV1

See [Bcp47ToIanaMap]

Constants

NON_REGION_CITY_PREFIX

IanaToBcp47Map’s trie cannot handle differently-cased prefixes, like Mexico/BajaSur`` and MET`.