icu::time::provider

Module windows

A provider for mapping Windows Zones to IANA identifiers.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Structs

TimezoneIdentifiersWindowsV1

See WindowsZonesToBcp47Map.

WindowsZonesToBcp47Map

A mapping from Windows Timezone names to the corresponding BCP-47 IDs.