icu::time::zone::models

Trait TimeZoneModel

pub trait TimeZoneModel: Sealed {
    type TimeZoneVariant: IntoOption<TimeZoneVariant> + Debug + Copy;
    type ZoneNameTimestamp: IntoOption<ZoneNameTimestamp> + Debug + Copy;
}

Trait encoding a particular data model for time zones.

๐ซ This trait is sealed; it cannot be implemented by user code. If an API requests an item that implements this trait, please consider using a type from the implementors listed below.

Required Associated Types

type TimeZoneVariant: IntoOption<TimeZoneVariant> + Debug + Copy

The zone variant, if required for this time zone model.

type ZoneNameTimestamp: IntoOption<ZoneNameTimestamp> + Debug + Copy

The local time, if required for this time zone model.