Struct icu::timezone::provider::names::Bcp47ToIanaMapV1

pub struct Bcp47ToIanaMapV1<'data> {
    pub bcp47_ids_checksum: u64,
    pub canonical_iana_ids: VarZeroVec<'data, str>,
}

A mapping from IANA time zone identifiers to BCP-47 time zone identifiers.

The BCP-47 time zone ID maps to the default IANA time zone ID according to the CLDR data.

๐ง This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Fields
bcp47_ids_checksum: u64

An XxHash64 checksum of IanaToBcp47MapV1::bcp47_ids.

The checksum here should match the checksum in IanaToBcp47MapV1 if these were generated from the same data set.

canonical_iana_ids: VarZeroVec<'data, str>

The IANA time zone identifier corresponding to the BCP-47 ID in IanaToBcp47MapV1::bcp47_ids.

Since there can be more than one IANA identifier for a particular BCP-47 identifier, this list contains only the current canonical IANA identifier.