icu::timezone
Struct IanaToBcp47Mapper

pub struct IanaToBcp47Mapper { /* private fields */ }

๐Deprecated since 1.5.0: use TimeZoneIdMapper instead

A mapper from IANA time zone identifiers to BCP-47 time zone identifiers.

Examples

use icu::timezone::IanaToBcp47Mapper;

let mapper = IanaToBcp47Mapper::new();
let mapper_borrowed = mapper.as_borrowed();

// The IANA zone "Australia/Melbourne" is the BCP-47 zone "aumel"
assert_eq!(
    mapper_borrowed.get("Australia/Melbourne"),
    Some("aumel".parse().unwrap())
);

// Lookup is ASCII-case insensitive
assert_eq!(
    mapper_borrowed.get("australia/melbourne"),
    Some("aumel".parse().unwrap())
);

// The IANA zone "Australia/Victoria" is an alias
assert_eq!(
    mapper_borrowed.get("Australia/Victoria"),
    Some("aumel".parse().unwrap())
);

Implementations

impl IanaToBcp47Mapper

pub const fn new() -> IanaToBcp47Mapper

Creates a new IanaToBcp47Mapper using compiled data.

See IanaToBcp47Mapper for an example.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

pub fn try_new_with_any_provider( provider: &(impl AnyProvider + ?Sized), ) -> Result<IanaToBcp47Mapper, TimeZoneError>

A version of [Self :: new] that uses custom data provided by an AnyProvider.

๐ Help choosing a constructor

pub fn try_new_with_buffer_provider( provider: &(impl BufferProvider + ?Sized), ) -> Result<IanaToBcp47Mapper, TimeZoneError>

A version of [Self :: new] that uses custom data provided by a BufferProvider.

โจ Enabled with the serde feature.

๐ Help choosing a constructor

pub fn try_new_unstable<P>( provider: &P, ) -> Result<IanaToBcp47Mapper, TimeZoneError>
where P: DataProvider<IanaToBcp47MapV1Marker> + ?Sized,

A version of Self::new that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

โ๏ธ The bounds on provider may change over time, including in SemVer minor releases.

pub fn as_borrowed(&self) -> IanaToBcp47MapperBorrowed<'_>

Returns a borrowed version of the mapper that can be queried.

This avoids a small potential cost of reading the data pointer.