Module icu::collator::provider

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Structs

Baked

Baked data

CollationDataV1

The main collation data either for the root or for a tailoring

CollationDataV1Marker

Marker type for CollationDataV1: “collator/data@1”

CollationDiacriticsV1

Secondary weights for the start of the Combining Diacritics block.

CollationDiacriticsV1Marker

Marker type for CollationDiacriticsV1: “collator/dia@1”

CollationJamoV1

CollationElement32s for the Hangul Jamo Unicode Block

CollationJamoV1Marker

Marker type for CollationJamoV1: “collator/jamo@1”

CollationMetadataV1

Each non-alias collation that the data provider knows about explicitly has an data entry at least for this struct.

CollationMetadataV1Marker

Marker type for CollationMetadataV1: “collator/meta@1”

CollationReorderingV1

Script reordering data

CollationReorderingV1Marker

Marker type for CollationReorderingV1: “collator/reord@1”

CollationSpecialPrimariesV1

Special primaries associated with the root collation

CollationSpecialPrimariesV1Marker

Marker type for CollationSpecialPrimariesV1: “collator/prim@1”