Struct icu::collator::Collator

pub struct Collator { /* private fields */ }

Compares strings according to culturally-relevant ordering.

Implementations

impl Collator

pub fn try_new( locale: &DataLocale, options: CollatorOptions ) -> Result<Collator, CollatorError>

Creates a collator for the given locale and options from compiled data.

pub fn try_new_with_any_provider( provider: &(impl AnyProvider + ?Sized), locale: &DataLocale, options: CollatorOptions ) -> Result<Collator, CollatorError>

A version of Self::try_new that uses custom data provided by an AnyProvider.

๐ Help choosing a constructor

pub fn try_new_with_buffer_provider( provider: &(impl BufferProvider + ?Sized), locale: &DataLocale, options: CollatorOptions ) -> Result<Collator, CollatorError>

A version of Self::try_new that uses custom data provided by a BufferProvider.

โจ Enabled with the serde feature.

๐ Help choosing a constructor

pub fn try_new_unstable<D>( provider: &D, locale: &DataLocale, options: CollatorOptions ) -> Result<Collator, CollatorError>where D: DataProvider<CollationSpecialPrimariesV1Marker> + DataProvider<CollationDataV1Marker> + DataProvider<CollationDiacriticsV1Marker> + DataProvider<CollationJamoV1Marker> + DataProvider<CollationMetadataV1Marker> + DataProvider<CollationReorderingV1Marker> + DataProvider<CanonicalDecompositionDataV1Marker> + DataProvider<CanonicalDecompositionTablesV1Marker> + ?Sized,

A version of Self::try_new that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

โ๏ธ The bounds on provider may change over time, including in SemVer minor releases.

pub fn compare_utf16(&self, left: &[u16], right: &[u16]) -> Ordering

Compare potentially ill-formed UTF-16 slices. Unpaired surrogates are compared as if each one was a REPLACEMENT CHARACTER.

pub fn compare(&self, left: &str, right: &str) -> Ordering

Compare guaranteed well-formed UTF-8 slices.

pub fn compare_utf8(&self, left: &[u8], right: &[u8]) -> Ordering

Compare potentially well-formed UTF-8 slices. Ill-formed input is compared as if errors had been replaced with REPLACEMENT CHARACTERs according to the WHATWG Encoding Standard.