Crate icu::datetime

Formatting date and time.

This module is published as its own crate (icu_datetime) and as part of the icu crate. See the latter for more details on the ICU4X project.

DateTimeFormatter is the main structure of the component. It accepts a set of arguments which allow it to collect necessary data from the data provider, and once instantiated, can be used to quickly format any date and time provided.

Examples

use icu::calendar::Gregorian;
use icu::datetime::{
    mock::parse_gregorian_from_str, options::length, DateTimeFormatter, DateTimeFormatterOptions,
};
use icu::locid::locale;

let provider = icu_testdata::get_provider();

// See the next code example for a more ergonomic example with .into().
let options = DateTimeFormatterOptions::Length(length::Bag::from_date_time_style(
    length::Date::Medium,
    length::Time::Short,
));

let dtf = DateTimeFormatter::<Gregorian>::try_new(&locale!("en").into(), &provider, &options)
    .expect("Failed to create DateTimeFormatter instance.");

let date = parse_gregorian_from_str("2020-09-12T12:35:00").expect("Failed to parse date.");

let formatted_date = dtf.format(&date);
assert_eq!(formatted_date.to_string(), "Sep 12, 2020, 12:35 PM");

The options can be created more ergonomically using the Into trait to automatically convert a options::length::Bag into a DateTimeFormatterOptions::Length.

use icu::calendar::Gregorian;
use icu::datetime::{options::length, DateTimeFormatter, DateTimeFormatterOptions};
let options =
    length::Bag::from_date_time_style(length::Date::Medium, length::Time::Short).into();

let dtf = DateTimeFormatter::<Gregorian>::try_new(&locale.into(), &provider, &options);

At the moment, the crate provides only options using the Length bag, but in the future, we expect to add more ways to customize the output, like skeletons, and components.

Notice: Rust at the moment does not have a canonical way to represent date and time. We use DateTime as an example of the data necessary for ICU DateTimeFormatter to work, and we hope to work with the community to develop core date and time APIs that will work as an input for this component. DateTime additionally has support for non-Gregorian calendars, which this module will eventually be able to format.

Modules

any

This module contains the untyped AnyCalendar-based DateTimeFormatter APIs that are capable of formatting dates from any calendar

date

A collection of utilities for representing and working with dates as an input to formatting operations.

datetime

The collection of code that is needed for handling formatting operations for DateTimes. Central to this is the DateTimeFormatter.

fields

Enums representing the fields in a date pattern, including the field’s type, length and symbol.

metazone

A module for metazone id calculation.

mock

A collection of temporary structs and utilities to input data for tests, benchmarks, and examples.

options

DateTimeFormatterOptions is a bag of options which, together with Locale, defines how dates will be formatted with a DateTimeFormatter instance.

provider

Data provider struct definitions for this ICU4X component.

time_zone

zoned_datetime

Structs

DateFormatter

DateFormatter is a structure of the icu_datetime component that provides date formatting only. When constructed, it uses data from the data provider, selected locale and provided preferences to collect all data necessary to format any date into that locale.

DateTimeFormatter

DateTimeFormatter is the main structure of the icu_datetime component. When constructed, it uses data from the data provider, selected locale and provided options to collect all data necessary to format any dates into that locale.

FormattedDateTime

FormattedDateTime is a intermediate structure which can be retrieved as an output from DateTimeFormatter.

FormattedTimeZone

FormattedTimeZone is a intermediate structure which can be retrieved as an output from TimeZoneFormatter.

FormattedZonedDateTime

TimeFormatter

TimeFormatter is a structure of the icu_datetime component that provides time formatting only. When constructed, it uses data from the data provider, selected locale and provided preferences to collect all data necessary to format any time into that locale.

TimeZoneFormatter

TimeZoneFormatter uses data from the data provider, the selected Locale, and the provided TimeZoneFormatterConfig to collect all data necessary to format time zones into that locale.

TimeZoneFormatterOptions

A bag of options to define how time zone will be formatted.

ZonedDateTimeFormatter

The composition of DateTimeFormatter and TimeZoneFormatter.

Enums

DateTimeFormatterError

A list of possible error outcomes for the DateTimeFormatter struct.

DateTimeFormatterOptions

A bag of options which, together with Locale, defines how dates will be formatted with a DateTimeFormatter instance.

TimeZoneFormatterConfig

A config enum for initializing TimeZoneFormatter.

Traits

CldrCalendar

A calendar that can be found in CLDR