Module icu::datetime

icu-datetime is one of the ICU4X components.

It is a core API for formatting date and time to user readable textual representation.

DateTimeFormat is the main structure of the component. It accepts a set of arguments which allow it to collect necessary data from the DataProvider, and once instantiated, can be used to quickly format any date and time provided.

Examples

use icu_locid_macros::langid;
use icu_datetime::{DateTimeFormat, options::style};
use icu_datetime::MockDateTime;
use icu_data_provider::InvariantDataProvider;

let lid = langid!("en");

let provider = InvariantDataProvider;

let options = style::Bag {
    date: Some(style::Date::Medium),
    time: Some(style::Time::Short),
    ..Default::default()
};
let dtf = DateTimeFormat::try_new(lid, &provider, &options.into())
    .expect("Failed to create DateTimeFormat instance.");

let date_time = MockDateTime::try_new(2020, 9, 1, 12, 34, 28)
    .expect("Failed to construct DateTime.");

let value = dtf.format_to_string(&date_time);

At the moment, the crate provides only options using the Style bag, but in the future, we expect to add more ways to customize the output, like skeletons, and components.

Notice: Rust at the moment does not have a canonical way to represent date and time. We are introducing date::MockDateTime as a reference example of the data necessary for ICU DateTimeFormat to work, and we hope to work with the community to develop core date and time APIs that will work as an input for this component.

Modules

date

options
DateTimeFormatOptions is a bag of options which, together with LanguageIdentifier, define how dates will be formatted be a DateTimeFormat instance.

Structs

DateTimeFormat
DateTimeFormat is the main structure of the icu-datetime component. When constructed, it uses data from the DataProvider, selected LanguageIdentifier and provided options to collect all data necessary to format any dates into that locale.

FormattedDateTime
FormattedDateTime is a intermediate structure which can be retrieved as an output from DateTimeFormat.

MockDateTime
Temporary implementation of DateTimeType, which is used in tests, benchmarks and examples of this component.

Enums

DateTimeFormatError
A list of possible error outcomes for the DateTimeFormat struct.

DateTimeFormatOptions
DateTimeFormatOptions is a bag of options which, together with LanguageIdentifier, define how dates will be formatted be a DateTimeFormat instance.

Traits

DateTimeType
Temporary trait used to represent the input data for DateTimeFormat.