Module icu::datetime::date

A collection of utilities for representing and working with dates as an input to formatting operations.

Structs

DayOfMonth

A day number in a month. Usually 1-based.

DayOfYearInfo

A struct containing various details about the position of the day within a year. It is returned

Era

TODO(#486): Implement era codes.

GmtOffset

The GMT offset in seconds for a MockTimeZone.

IsoHour

An ISO-8601 hour component, for use with the [IsoTimeInput].

IsoMinute

An ISO-8601 minute component, for use with the [IsoTimeInput].

IsoSecond

An ISO-8601 second component, for use with the [IsoTimeInput].

Month

Representation of a formattable month.

MonthCode

TODO(#486): Implement month codes.

Time

WeekOfMonth

A week number in a month. Usually 1-based.

WeekOfYear

A week number in a year. Usually 1-based.

Year

Representation of a formattable year.

Enums

DateTimeError

A list of possible error outcomes for working with various inputs to DateTime inputs and operations.

FractionalSecond

A placeholder for fractional seconds support. See Issue #485 for tracking the support of this feature.

IsoWeekday

A weekday in a 7-day week, according to ISO-8601.

Traits

DateInput

Representation of a formattable calendar date. Supports dates in any calendar system that uses solar days indexed by an era, year, month, and day.

DateTimeInput

A combination of a formattable calendar date and ISO time.

IsoTimeInput

Representation of a time of day according to ISO-8601 conventions. Always indexed from midnight, regardless of calendar system.

LocalizedDateTimeInput

A formattable calendar date and ISO time that takes the locale into account.

TimeZoneInput

Representation of a formattable time zone.

ZonedDateTimeInput

A combination of a formattable calendar date, ISO time, and time zone.