icu::datetime
Module fields

Enums representing the fields in a date pattern, including the field’s type, length and symbol.

Structs

DayPeriodULE

ULE type for DayPeriod

DayULE

ULE type for Day

Field

A field within a date pattern string, also referred to as a date field. A date field is the repetition of a specific pattern character one or more times within the pattern string. The pattern character is known as the field symbol, which indicates the particular meaning for the field.

FieldSymbolULE

ULE type for FieldSymbol

FieldULE

ULE type for Field

HourULE

ULE type for Hour

MonthULE

ULE type for Month

SecondULE

ULE type for Second

TimeZoneULE

ULE type for TimeZone

WeekULE

ULE type for Week

WeekdayULE

ULE type for Weekday

YearULE

ULE type for Year

Enums

Day

An enum for the possible symbols of a day field in a date pattern.

DayPeriod

An enum for the possible symbols of a day period field in a date pattern.

Error

An error relating to the field for a date pattern field as a whole.

FieldLength

An enum representing the length of a field within a date or time formatting pattern string, in which the pattern field is represented as a letter occurring 1 or more times in a row, ex: MMM, dd, y. See the LDML documentation in UTS 35 for more details.

FieldNumericOverrides

Various numeric overrides for datetime patterns as found in CLDR

FieldSymbol

A field symbol for a date formatting pattern. Field symbols are a more granular distinction for a pattern field within the category of a field type. Examples of field types are: Year, Month, Hour. Within the Hour field type, examples of field symbols are: Hour::H12, Hour::H24. Each field symbol is represented within the date formatting pattern string by a distinct character from the set of A..Z and a..z.

Hour

An enum for the possible symbols of an hour field in a date pattern.

LengthError

An error relating to the length of a field within a date pattern.

Month

An enum for the possible symbols of a month field in a date pattern.

Second

An enum for the possible symbols of a second field in a date pattern.

SymbolError

An error relating to the field symbol for a date pattern field.

TimeZone

An enum for the possible symbols of a time zone field in a date pattern.

Week

An enum for the possible symbols of a week field in a date pattern.

Weekday

An enum for the possible symbols of a weekday field in a date pattern.

Year

An enum for the possible symbols of a year field in a date pattern.