Trait icu::datetime::input::IsoTimeInput

pub trait IsoTimeInput {
    // Required methods
    fn hour(&self) -> Option<IsoHour>;
    fn minute(&self) -> Option<IsoMinute>;
    fn second(&self) -> Option<IsoSecond>;
    fn nanosecond(&self) -> Option<NanoSecond>;
}

Representation of a time of day according to ISO-8601 conventions. Always indexed from midnight, regardless of calendar system.

All fields are optional. If a field is not present but is required when formatting, an error result will be returned from the formatter.

All data represented in IsoTimeInput should be locale-agnostic.

Required Methods

fn hour(&self) -> Option<IsoHour>

Gets the hour input.

fn minute(&self) -> Option<IsoMinute>

Gets the minute input.

fn second(&self) -> Option<IsoSecond>

Gets the second input.

fn nanosecond(&self) -> Option<NanoSecond>

Gets the nanosecond input.