Struct icu::datetime::FormattedDateTime

pub struct FormattedDateTime<'l, T> where
    T: DateTimeInput,  { /* fields omitted */ }

FormattedDateTime is a intermediate structure which can be retrieved as an output from DateTimeFormat.

The structure contains all the information needed to display formatted value, and it will also contain additional methods allowing the user to introspect and even manipulate the formatted data.

Examples

use icu::locid::Locale;
use icu::locid::macros::langid;
use icu::datetime::{DateTimeFormat, DateTimeFormatOptions};
use icu::datetime::mock::datetime::MockDateTime;
use icu_provider::inv::InvariantDataProvider;
let locale: Locale = langid!("en").into();
let dtf = DateTimeFormat::try_new(locale, &provider, &options)
    .expect("Failed to create DateTimeFormat instance.");

let datetime = MockDateTime::try_new(2020, 9, 1, 12, 34, 28)
    .expect("Failed to construct DateTime.");

let formatted_date = dtf.format(&datetime);

let _ = format!("Date: {}", formatted_date);