icu::datetime
Struct TypedZonedDateTimeFormatter

pub struct TypedZonedDateTimeFormatter<C>(/* private fields */);

The composition of TypedDateTimeFormatter and TimeZoneFormatter.

TypedZonedDateTimeFormatter is a formatter capable of formatting date/times with time zones from a calendar selected at compile time. For the difference between this and DateTimeFormatter, please read the crate root docs.

TypedZonedDateTimeFormatter uses data from the [data provider]s, the selected locale, and the provided pattern to collect all data necessary to format a datetime with time zones into that locale.

The various pattern symbols specified in UTS-35 require different sets of data for formatting. As such, TimeZoneFormatter will pull in only the resources it needs to format that pattern that is derived from the provided DateTimeFormatterOptions.

For that reason, one should think of the process of formatting a zoned datetime in two steps: first, a computationally heavy construction of TypedZonedDateTimeFormatter, and then fast formatting of the data using the instance.

Examples

use icu::calendar::{DateTime, Gregorian};
use icu::datetime::time_zone::TimeZoneFormatterOptions;
use icu::datetime::{options::length, TypedZonedDateTimeFormatter};
use icu::locid::locale;
use icu::timezone::CustomTimeZone;
use std::str::FromStr;
use writeable::assert_writeable_eq;

let options = length::Bag::from_date_time_style(
    length::Date::Medium,
    length::Time::Long,
);
let zdtf = TypedZonedDateTimeFormatter::<Gregorian>::try_new(
    &locale!("en").into(),
    options.into(),
    TimeZoneFormatterOptions::default(),
)
.expect("Failed to create TypedDateTimeFormatter instance.");

let datetime =
    DateTime::try_new_gregorian_datetime(2020, 9, 12, 12, 34, 28).unwrap();
let time_zone = CustomTimeZone::from_str("-07:00").unwrap();

let formatted_date = zdtf.format(&datetime, &time_zone);

assert_writeable_eq!(formatted_date, "Sep 12, 2020, 12:34:28โฏPM GMT-07:00");

Implementations

impl<C> TypedZonedDateTimeFormatter<C>
where C: CldrCalendar,

pub fn try_new( locale: &DataLocale, date_time_format_options: DateTimeFormatterOptions, time_zone_format_options: TimeZoneFormatterOptions, ) -> Result<TypedZonedDateTimeFormatter<C>, DateTimeError>
where Baked: DataProvider<<C as CldrCalendar>::DateLengthsV1Marker> + DataProvider<<C as CldrCalendar>::DateSymbolsV1Marker>,

Constructor that takes a selected locale and a list of DateTimeFormatterOptions. It collects all data necessary to format zoned datetime values into the given locale.

โจ Enabled with the compiled_data Cargo feature.

๐ Help choosing a constructor

Examples

use icu::calendar::{DateTime, Gregorian};
use icu::datetime::time_zone::TimeZoneFormatterOptions;
use icu::datetime::{options::length, TypedZonedDateTimeFormatter};
use icu::locid::locale;
use icu::timezone::CustomTimeZone;
use writeable::assert_writeable_eq;

let options = length::Bag::from_date_time_style(
    length::Date::Medium,
    length::Time::Long,
);

let zdtf = TypedZonedDateTimeFormatter::<Gregorian>::try_new(
    &locale!("en").into(),
    options.into(),
    TimeZoneFormatterOptions::default(),
)
.unwrap();

let datetime =
    DateTime::try_new_gregorian_datetime(2022, 8, 31, 1, 2, 3).unwrap();

assert_writeable_eq!(
    zdtf.format(&datetime, &CustomTimeZone::utc()),
    "Aug 31, 2022, 1:02:03โฏAM GMT",
);

pub fn try_new_with_any_provider( provider: &(impl AnyProvider + ?Sized), locale: &DataLocale, date_time_format_options: DateTimeFormatterOptions, time_zone_format_options: TimeZoneFormatterOptions, ) -> Result<TypedZonedDateTimeFormatter<C>, DateTimeError>

A version of [Self :: try_new] that uses custom data provided by an AnyProvider.

๐ Help choosing a constructor

pub fn try_new_with_buffer_provider( provider: &(impl BufferProvider + ?Sized), locale: &DataLocale, date_time_format_options: DateTimeFormatterOptions, time_zone_format_options: TimeZoneFormatterOptions, ) -> Result<TypedZonedDateTimeFormatter<C>, DateTimeError>

A version of [Self :: try_new] that uses custom data provided by a BufferProvider.

โจ Enabled with the serde feature.

๐ Help choosing a constructor

pub fn try_new_unstable<P>( provider: &P, locale: &DataLocale, date_time_format_options: DateTimeFormatterOptions, time_zone_format_options: TimeZoneFormatterOptions, ) -> Result<TypedZonedDateTimeFormatter<C>, DateTimeError>
where P: DataProvider<<C as CldrCalendar>::DateSymbolsV1Marker> + DataProvider<<C as CldrCalendar>::DateLengthsV1Marker> + DataProvider<TimeSymbolsV1Marker> + DataProvider<TimeLengthsV1Marker> + DataProvider<WeekDataV1Marker> + DataProvider<TimeZoneFormatsV1Marker> + DataProvider<ExemplarCitiesV1Marker> + DataProvider<MetazoneGenericNamesLongV1Marker> + DataProvider<MetazoneGenericNamesShortV1Marker> + DataProvider<MetazoneSpecificNamesLongV1Marker> + DataProvider<MetazoneSpecificNamesShortV1Marker> + DataProvider<OrdinalV1Marker> + DataProvider<DecimalSymbolsV1Marker> + ?Sized,

A version of Self::try_new that uses custom data provided by a DataProvider.

๐ Help choosing a constructor

โ๏ธ The bounds on provider may change over time, including in SemVer minor releases.

pub fn format<'l>( &'l self, date: &impl DateTimeInput<Calendar = C>, time_zone: &impl TimeZoneInput, ) -> FormattedZonedDateTime<'l>

Takes a DateTimeInput and a TimeZoneInput and returns an instance of a FormattedZonedDateTime that contains all information necessary to display a formatted zoned datetime and operate on it.

Examples

use icu::calendar::{DateTime, Gregorian};
use icu::datetime::{options::length, TypedZonedDateTimeFormatter};
use icu::locid::locale;
use icu::timezone::CustomTimeZone;
use std::str::FromStr;
use writeable::assert_writeable_eq;

let options = length::Bag::from_date_time_style(
    length::Date::Medium,
    length::Time::Long,
);

let zdtf = TypedZonedDateTimeFormatter::<Gregorian>::try_new(
    &locale!("en").into(),
    options.into(),
    Default::default(),
)
.expect("Failed to create TypedZonedDateTimeFormatter instance.");

let datetime =
    DateTime::try_new_gregorian_datetime(2020, 9, 12, 12, 34, 28).unwrap();
let time_zone = CustomTimeZone::from_str("-07:00").unwrap();

let formatted_date = zdtf.format(&datetime, &time_zone);

assert_writeable_eq!(formatted_date, "Sep 12, 2020, 12:34:28โฏPM GMT-07:00");

pub fn format_to_string( &self, date: &impl DateTimeInput<Calendar = C>, time_zone: &impl TimeZoneInput, ) -> String

Takes a DateTimeInput and a TimeZoneInput and returns it formatted as a string.

Examples

use icu::calendar::{DateTime, Gregorian};
use icu::datetime::{options::length, TypedZonedDateTimeFormatter};
use icu::locid::locale;
use icu::timezone::CustomTimeZone;
use std::str::FromStr;

let options = length::Bag::from_date_time_style(
    length::Date::Medium,
    length::Time::Long,
);

let zdtf = TypedZonedDateTimeFormatter::<Gregorian>::try_new(
    &locale!("en").into(),
    options.into(),
    Default::default(),
)
.expect("Failed to create TypedZonedDateTimeFormatter instance.");

let datetime =
    DateTime::try_new_gregorian_datetime(2020, 9, 12, 12, 34, 28).unwrap();
let time_zone = CustomTimeZone::from_str("-07:00").unwrap();

let formatted_string = zdtf.format_to_string(&datetime, &time_zone);

assert_eq!(formatted_string, "Sep 12, 2020, 12:34:28โฏPM GMT-07:00");