Crate icu::decimal

Formatting basic decimal numbers.

This module is published as its own crate (icu_decimal) and as part of the icu crate. See the latter for more details on the ICU4X project.

Support for currencies, measurement units, and compact notation is planned. To track progress, follow icu4x#275.

Examples

Format a number with Bangla digits

use fixed_decimal::FixedDecimal;
use icu::decimal::FixedDecimalFormatter;
use icu::locid::locale;
use writeable::assert_writeable_eq;

let fdf = FixedDecimalFormatter::try_new_unstable(
    &icu_testdata::unstable(),
    &locale!("bn").into(),
    Default::default(),
)
.expect("Data should load successfully");

let fixed_decimal = FixedDecimal::from(1000007);

assert_writeable_eq!(fdf.format(&fixed_decimal), "১০,০০,০০৭");

Format a number with digits after the decimal separator

use fixed_decimal::FixedDecimal;
use icu::decimal::FixedDecimalFormatter;
use icu::locid::Locale;
use writeable::assert_writeable_eq;

let fdf = FixedDecimalFormatter::try_new_unstable(
    &icu_testdata::unstable(),
    &Locale::UND.into(),
    Default::default(),
)
.expect("Data should load successfully");

let fixed_decimal = FixedDecimal::from(200050).multiplied_pow10(-2);

assert_writeable_eq!(fdf.format(&fixed_decimal), "2,000.50");

Format a number using an alternative numbering system

Numbering systems specified in the -u-nu subtag will be followed as long as the locale has symbols for that numbering system.

use fixed_decimal::FixedDecimal;
use icu::decimal::FixedDecimalFormatter;
use icu::locid::locale;
use writeable::assert_writeable_eq;

let fdf = FixedDecimalFormatter::try_new_unstable(
    &icu_testdata::unstable(),
    &locale!("th-u-nu-thai").into(),
    Default::default(),
)
.expect("Data should load successfully");

let fixed_decimal = FixedDecimal::from(1000007);

assert_writeable_eq!(fdf.format(&fixed_decimal), "๑,๐๐๐,๐๐๗");

Modules

options

Options for FixedDecimalFormatter.

provider

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

Structs

FixedDecimalFormatter

A formatter for FixedDecimal, rendering decimal digits in an i18n-friendly way.

FormattedFixedDecimal

An intermediate structure returned by FixedDecimalFormatter. Use Writeable to render the formatted decimal to a string or buffer.

Enums

DecimalError

A list of error outcomes for various operations in this module.

Error

A list of error outcomes for various operations in this module.