icu::decimal
Module provider

🚧 [Unstable] Data provider struct definitions for this ICU4X component.

🚧 This code is considered unstable; it may change at any time, in breaking or non-breaking ways, including in SemVer minor releases. While the serde representation of data structs is guaranteed to be stable, their Rust representation might not be. Use with caution.

Read more about data providers: icu_provider

Structs

AffixesV1

A collection of strings to affix to a decimal number.

Baked

Baked data

DecimalSymbolsV1

Symbols and metadata required for formatting a FixedDecimal.

DecimalSymbolsV1Marker

Marker type for DecimalSymbolsV1: “decimal/symbols@1”

GroupingSizesV1

A collection of settings expressing where to put grouping separators in a decimal number. For example, 1,000,000 has two grouping separators, positioned along every 3 digits.