Module icu::locid

Language and Locale operations

This API provides necessary functionality for parsing, manipulating, and serializing Unicode Language and Locale Identifiers.

The crate provides algorithms for parsing a string into a well-formed language or locale identifier as defined by UTS #35: Unicode LDML 3. Unicode Language and Locale Identifiers.

Locale is the most common structure to use for storing information about a language, script, region, variants and extensions. In almost all cases, this struct should be used as the base unit for all locale management operations.

LanguageIdentifier is a strict subset of Locale which can be useful in a narrow range of cases where Unicode Extensions are not relevant.

If in doubt, use Locale.

Examples

use icu::locid::Locale;
use icu::locid::subtags::{Language, Region};

let mut loc: Locale = "en-US".parse()
    .expect("Parsing failed.");

let lang: Language = "en".parse()
    .expect("Parsing failed.");
let region: Region = "US".parse()
    .expect("Parsing failed.");

assert_eq!(loc.id.language, lang);
assert_eq!(loc.id.script, None);
assert_eq!(loc.id.region, Some(region));
assert_eq!(loc.id.variants.len(), 0);

let region: Region = "GB".parse().expect("Parsing failed.");
loc.id.region = Some(region);

assert_eq!(loc.to_string(), "en-GB");

For more details, see Locale and LanguageIdentifier.

Modules

extensions
Unicode Extensions provide a mechanism to extend the LanguageIdentifier with additional bits of information - a combination of a LanguageIdentifier and Extensions is called Locale.

macros

subtags
Language Identifier and Locale contains a set of subtags which represent different fields of the structure.

Structs

LanguageIdentifier
LanguageIdentifier is a core struct representing a Unicode BCP47 Language Identifier.

Locale
Locale is a core struct representing a Unicode Locale Identifier.

Enums

ParserError
List of parser errors that can be generated while parsing LanguageIdentifier, Locale, subtags or extensions.