Enum icu::properties::GeneralCategory

pub enum GeneralCategory {

    Unassigned,
    UppercaseLetter,
    LowercaseLetter,
    TitlecaseLetter,
    ModifierLetter,
    OtherLetter,
    NonspacingMark,
    SpacingMark,
    EnclosingMark,
    DecimalNumber,
    LetterNumber,
    OtherNumber,
    SpaceSeparator,
    LineSeparator,
    ParagraphSeparator,
    Control,
    Format,
    PrivateUse,
    Surrogate,
    DashPunctuation,
    OpenPunctuation,
    ClosePunctuation,
    ConnectorPunctuation,
    InitialPunctuation,
    FinalPunctuation,
    OtherPunctuation,
    MathSymbol,
    CurrencySymbol,
    ModifierSymbol,
    OtherSymbol,

}

Enumerated property General_Category.

General_Category specifies the most general classification of a code point, usually determined based on the primary characteristic of the assigned character. For example, is the character a letter, a mark, a number, punctuation, or a symbol, and if so, of what type?

GeneralCategory only supports specific subcategories (eg UppercaseLetter). It does not support grouped categories (eg Letter). For grouped categories, use GeneralCategoryGroup.

Variants

Unassigned

(Cn) A reserved unassigned code point or a noncharacter

UppercaseLetter

(Lu) An uppercase letter

LowercaseLetter

(Ll) A lowercase letter

TitlecaseLetter

(Lt) A digraphic letter, with first part uppercase

ModifierLetter

(Lm) A modifier letter

OtherLetter

(Lo) Other letters, including syllables and ideographs

NonspacingMark

(Mn) A nonspacing combining mark (zero advance width)

SpacingMark

(Mc) A spacing combining mark (positive advance width)

EnclosingMark

(Me) An enclosing combining mark

DecimalNumber

(Nd) A decimal digit

LetterNumber

(Nl) A letterlike numeric character

OtherNumber

(No) A numeric character of other type

SpaceSeparator

(Zs) A space character (of various non-zero widths)

LineSeparator

(Zl) U+2028 LINE SEPARATOR only

ParagraphSeparator

(Zp) U+2029 PARAGRAPH SEPARATOR only

Control

(Cc) A C0 or C1 control code

Format

(Cf) A format control character

PrivateUse

(Co) A private-use character

Surrogate

(Cs) A surrogate code point

DashPunctuation

(Pd) A dash or hyphen punctuation mark

OpenPunctuation

(Ps) An opening punctuation mark (of a pair)

ClosePunctuation

(Pe) A closing punctuation mark (of a pair)

ConnectorPunctuation

(Pc) A connecting punctuation mark, like a tie

InitialPunctuation

(Pi) An initial quotation mark

FinalPunctuation

(Pf) A final quotation mark

OtherPunctuation

(Po) A punctuation mark of other type

MathSymbol

(Sm) A symbol of mathematical use

CurrencySymbol

(Sc) A currency sign

ModifierSymbol

(Sk) A non-letterlike modifier symbol

OtherSymbol

(So) A symbol of other type

Implementations

impl GeneralCategory

pub fn get_name_to_enum_mapper( provider: &impl DataProvider<GeneralCategoryNameToValueV1Marker> ) -> Result<PropertyValueNameToEnumMapper<GeneralCategory>, PropertiesError>

Return a PropertyValueNameToEnumMapper, capable of looking up values from strings for the General_Category enumerated property

Example

use icu::properties::GeneralCategory;

let lookup = GeneralCategory::get_name_to_enum_mapper(&icu_testdata::unstable())
                 .expect("The data should be valid");
let lookup = lookup.as_borrowed();
// short name for value
assert_eq!(lookup.get_strict("Lu"), Some(GeneralCategory::UppercaseLetter));
assert_eq!(lookup.get_strict("Pd"), Some(GeneralCategory::DashPunctuation));
// long name for value
assert_eq!(lookup.get_strict("Uppercase_Letter"), Some(GeneralCategory::UppercaseLetter));
assert_eq!(lookup.get_strict("Dash_Punctuation"), Some(GeneralCategory::DashPunctuation));
// name has incorrect casing
assert_eq!(lookup.get_strict("dashpunctuation"), None);
// loose matching of name
assert_eq!(lookup.get_loose("dash-punctuation"), Some(GeneralCategory::DashPunctuation));
// fake property
assert_eq!(lookup.get_loose("Animated_Gif"), None);

pub fn get_enum_to_short_name_mapper( provider: &impl DataProvider<GeneralCategoryValueToShortNameV1Marker> ) -> Result<PropertyEnumToValueNameLinearMapper<GeneralCategory>, PropertiesError>

Return a PropertyEnumToValueNameLinearMapper, capable of looking up short names for values of the General_Category enumerated property

Example

use icu::properties::GeneralCategory;

let lookup = GeneralCategory::get_enum_to_short_name_mapper(&icu_testdata::unstable())
                 .expect("The data should be valid");
let lookup = lookup.as_borrowed();
assert_eq!(lookup.get(GeneralCategory::UppercaseLetter), Some("Lu"));
assert_eq!(lookup.get(GeneralCategory::DashPunctuation), Some("Pd"));

pub fn get_enum_to_long_name_mapper( provider: &impl DataProvider<GeneralCategoryValueToLongNameV1Marker> ) -> Result<PropertyEnumToValueNameLinearMapper<GeneralCategory>, PropertiesError>

Return a PropertyEnumToValueNameLinearMapper, capable of looking up long names for values of the General_Category enumerated property

Example

use icu::properties::GeneralCategory;

let lookup = GeneralCategory::get_enum_to_long_name_mapper(&icu_testdata::unstable())
                 .expect("The data should be valid");
let lookup = lookup.as_borrowed();
assert_eq!(lookup.get(GeneralCategory::UppercaseLetter), Some("Uppercase_Letter"));
assert_eq!(lookup.get(GeneralCategory::DashPunctuation), Some("Dash_Punctuation"));