Crate icu::timezone

Types for resolving and manipulating timezones

Modules

provider

Data provider struct definitions for this ICU4X component.

Structs

CustomTimeZone

A utility type that can hold time zone information

GmtOffset

The GMT offset in seconds for a timezone

MetaZoneCalculator

MetaZoneCalculator uses data from the data provider to calculate metazone id.

ZoneVariant

A time zone variant: currently either daylight time or standard time.

Enums

TimeZoneError

A list of possible error outcomes for working with types in this crate and operations.